CardanoNode

Documentation

data NodeLog

Constructors

MsgNodeCmdSpec

Fields

cmd :: Text

MsgCLI [Text]

MsgCLIStatus Text Text

MsgCLIRetry Text

MsgCLIRetryResult Text Int

MsgNodeStarting

Fields

stateDirectory :: FilePath

MsgSocketIsReady SocketPath

MsgSynchronizing

Fields

percentDone :: Centi

timeDifference :: NominalDiffTime

blockTime :: NominalDiffTime

tipTime :: NominalDiffTime

targetTime :: NominalDiffTime

MsgQueryGenesisParametersFailed

Fields

err :: Text

type Port = Int

newtype NodeId

Constructors

NodeId Int

data DevnetConfig

Configuration parameters for a single node devnet

Constructors

DevnetConfig

Fields

stateDirectory :: FilePath

Parent state directory

systemStart :: UTCTime

Blockchain start time

ports :: PortsConfig

A list of port

data CardanoNodeArgs

Arguments given to the 'cardano-node' command-line to run a node.

Constructors

CardanoNodeArgs

Fields

nodeSocket :: FilePath

nodeConfigFile :: FilePath

nodeByronGenesisFile :: FilePath

nodeShelleyGenesisFile :: FilePath

nodeAlonzoGenesisFile :: FilePath

nodeConwayGenesisFile :: FilePath

nodeTopologyFile :: FilePath

nodeDatabaseDir :: FilePath

nodeDlgCertFile :: Maybe FilePath

nodeSignKeyFile :: Maybe FilePath

nodeOpCertFile :: Maybe FilePath

nodeKesKeyFile :: Maybe FilePath

nodeVrfKeyFile :: Maybe FilePath

nodePort :: Maybe Port

defaultCardanoNodeArgs :: CardanoNodeArgs

data PortsConfig

Configuration of ports from the perspective of a peer in the context of a fully sockected topology.

Constructors

PortsConfig

Fields

ours :: Port

Our node TCP port.

peers :: [Port]

Other peers TCP ports.

getCardanoNodeVersion :: IO String

findRunningCardanoNode :: Tracer IO NodeLog -> FilePath -> KnownNetwork -> IO (Maybe (NominalDiffTime, DirectBackend))

Tries to find an communicate with an existing cardano-node running in given work directory. NOTE: This is using the default node socket name as defined by defaultCardanoNodeArgs.

findRunningCardanoNode' :: Tracer IO NodeLog -> NetworkId -> SocketPath -> IO (Maybe (NominalDiffTime, DirectBackend))

Tries to find an communicate with an existing cardano-node running in given network id and socket path.

withCardanoNodeDevnet

Arguments

:: Tracer IO NodeLog

-> FilePath
State directory in which credentials, db & logs are persisted.

-> (NominalDiffTime -> DirectBackend -> IO a)

-> IO a

Start a single cardano-node devnet using the config from config/ and credentials from configcredentials. Only the Faucet actor will receive "initialFunds". Use seedFromFaucet to distribute funds other wallets.

withBlockfrostBackend

Arguments

:: Tracer IO NodeLog

-> FilePath
State directory in which credentials, db & logs are persisted.

-> (NominalDiffTime -> BlockfrostBackend -> IO a)

-> IO a

findFileStartingAtDirectory :: Int -> FilePath -> IO FilePath

Find the given file in the current directory or its parents.

This function starts from the current working directory and checks if the specified file exists there. If not found, it recursively checks the parent directories up to the given maximum depth.

data HydraBackend

Constructors

DirectBackendType

BlockfrostBackendType

getHydraBackend :: IO HydraBackend

withBackend :: forall a. Tracer IO NodeLog -> FilePath -> (forall backend. ChainBackend backend => NominalDiffTime -> backend -> IO a) -> IO a

withCardanoNodeOnKnownNetwork

Arguments

:: Tracer IO NodeLog

-> FilePath
State directory in which node db & logs are persisted.

-> KnownNetwork
A well-known Cardano network to connect to.

-> (NominalDiffTime -> DirectBackend -> IO a)

-> IO a

Run a cardano-node as normal network participant on a known network.

setupCardanoDevnet :: FilePath -> IO CardanoNodeArgs

Setup the cardano-node to run a local devnet producing blocks. This copies the appropriate files and prepares CardanoNodeArgs for withCardanoNode.

withCardanoNode :: Tracer IO NodeLog -> FilePath -> CardanoNodeArgs -> (NominalDiffTime -> DirectBackend -> IO a) -> IO a

getShelleyGenesisBlockTime :: Value -> NominalDiffTime

computeBlockTime :: NominalDiffTime -> Rational -> NominalDiffTime

Compute the block time (expected time between blocks) given a slot length as diff time and active slot coefficient.

waitForFullySynchronized :: ChainBackend backend => Tracer IO NodeLog -> backend -> IO ()

Wait until the node is fully caught up with the network. This can take a while!

waitForSocket :: SocketPath -> IO ()

Wait for the node socket file to become available.

cardanoNodeProcess :: Maybe FilePath -> CardanoNodeArgs -> CreateProcess

Generate command-line arguments for launching cardano-node.

initSystemStart :: IO UTCTime

Initialize the system start time to now (modulo a small offset needed to give time to the system to bootstrap correctly).

refreshSystemStart

Arguments

:: FilePath
Working directory in which paths of CardanoNodeArgs are resolved.

-> CardanoNodeArgs

-> IO ()

Re-generate configuration and genesis files with fresh system start times.

mkTopology :: [Port] -> Value

Generate a topology file from a list of peers.

data ProcessHasExited

Constructors

ProcessHasExited Text ExitCode

cliQueryProtocolParameters :: SocketPath -> NetworkId -> IO Value

Cardano-cli wrapper to query protocol parameters. While we have also client functions in Hydra.Chain.CardanoClient and Hydra.Cluster.CardanoClient, sometimes we deliberately want to use the cardano-cli to ensure compatibility.

unsafeDecodeJson :: FromJSON a => ByteString -> IO a

unsafeDecodeJsonFile :: FromJSON a => FilePath -> IO a