Data.EventSource

Description

The types and functions for creating EventSource and EventSink handles, intended to be used as an extension point.

A single EventSource and zero or more EventSink handles let an application load and emit an ordered, identified event stream. SQLite is a ready-made implementation.

Documentation

type EventId = Word64

class HasEventId a where

Methods

getEventId :: a -> EventId

newtype EventSource e m

Constructors

EventSource

Fields

sourceEvents :: HasEventId e => ConduitT () e (ResourceT m) ()

Stream all events from the event source.

getEvents :: (HasEventId e, MonadUnliftIO m) => EventSource e m -> m [e]

Retrieve all events from the event source as a list.

data EventSink e m

Constructors

EventSink

Fields

putEvent :: HasEventId e => e -> m ()

Send a single event to the event sink.

putEvents :: HasEventId e => [e] -> m ()

Send a batch of events to the event sink.

mkEventSink :: Monad m => (HasEventId e => e -> m ()) -> EventSink e m

Create an EventSink from a single-event function, with a default sequential batch implementation.

putEventsToSinks :: (Monad m, HasEventId e) => [EventSink e m] -> [e] -> m ()

Put a list of events to a list of event sinks, batching per sink.