Hydra.API.ServerOutput

Documentation

data TimedServerOutput tx

The type of messages sent to clients by the Server.

Constructors

TimedServerOutput

Fields

output :: ServerOutput tx

seq :: Natural

time :: UTCTime

data DecommitInvalidReason tx

Constructors

DecommitTxInvalid

Fields

localUTxO :: UTxOType tx

validationError :: ValidationError

DecommitAlreadyInFlight

Fields

otherDecommitTxId :: TxIdType tx

data ClientMessage tx

Individual messages as produced by the HeadLogic in the ClientEffect.

Constructors

CommandFailed

Fields

clientInput :: ClientInput tx

state :: HeadState tx

PostTxOnChainFailed

Fields

postChainTx :: PostChainTx tx

postTxError :: PostTxError tx

RejectedInput

Fields

clientInput :: ClientInput tx

reason :: Text

SideLoadSnapshotRejected

Fields

clientInput :: ClientInput tx

requirementFailure :: SideLoadRequirementFailure tx

data Greetings tx

A friendly welcome message which tells a client something about the node. Currently used for knowing what signing key the server uses (it only knows one), HeadStatus and optionally (if HeadIsOpen or SnapshotConfirmed message is emitted) UTxO's present in the Hydra Head.

Constructors

Greetings

Fields

me :: Party

headStatus :: HeadStatus

hydraHeadId :: Maybe HeadId

snapshotUtxo :: Maybe (UTxOType tx)

hydraNodeVersion :: String

env :: Environment

networkInfo :: NetworkInfo

chainSyncedStatus :: SyncedStatus

currentSlot :: ChainSlot

data InvalidInput

Constructors

InvalidInput

Fields

reason :: String

input :: Text

data ServerOutput tx

Constructors

NetworkConnected

NetworkDisconnected

NetworkVersionMismatch

Fields

ourVersion :: ProtocolVersion

theirVersion :: Maybe ProtocolVersion

NetworkClusterIDMismatch

Fields

clusterPeers :: Text

misconfiguredPeers :: Text

PeerConnected

Fields

peer :: Host

PeerDisconnected

Fields

peer :: Host

HeadIsInitializing

Fields

headId :: HeadId

parties :: [Party]

Committed

Fields

headId :: HeadId

party :: Party

utxo :: UTxOType tx

HeadIsOpen

Fields

headId :: HeadId

utxo :: UTxOType tx

HeadIsClosed

Fields

headId :: HeadId

snapshotNumber :: SnapshotNumber

contestationDeadline :: UTCTime

Nominal deadline until which contest can be submitted and after which fanout is possible. NOTE: Use this only for informational purpose and wait for ReadyToFanout instead before sending Fanout as the ledger of our cardano-node might not have progressed sufficiently in time yet and we do not re-submit transactions (yet).

HeadIsContested

Fields

headId :: HeadId

snapshotNumber :: SnapshotNumber

contestationDeadline :: UTCTime

Nominal deadline until which contest can be submitted and after which fanout is possible. NOTE: Use this only for informational purpose and wait for ReadyToFanout instead before sending Fanout as the ledger of our cardano-node might not have progressed sufficiently in time yet and we do not re-submit transactions (yet).

ReadyToFanout

Fields

headId :: HeadId

HeadIsAborted

Fields

headId :: HeadId

utxo :: UTxOType tx

HeadIsFinalized

Fields

headId :: HeadId

utxo :: UTxOType tx

TxValid
Given transaction has been seen as valid in the Head. It is expected to eventually be part of a SnapshotConfirmed.

Fields

headId :: HeadId

transactionId :: TxIdType tx

TxInvalid
Given transaction was not not applicable to the given UTxO in time and has been dropped.

Fields

headId :: HeadId

utxo :: UTxOType tx

transaction :: tx

validationError :: ValidationError

SnapshotConfirmed
Given snapshot was confirmed and included transactions can be considered final.

Fields

headId :: HeadId

snapshot :: Snapshot tx

signatures :: MultiSignature (Snapshot tx)

IgnoredHeadInitializing

Fields

headId :: HeadId

contestationPeriod :: ContestationPeriod

parties :: [Party]

participants :: [OnChainId]

DecommitRequested

Fields

headId :: HeadId

decommitTx :: tx

utxoToDecommit :: UTxOType tx

DecommitInvalid

Fields

headId :: HeadId

decommitTx :: tx

decommitInvalidReason :: DecommitInvalidReason tx

DecommitApproved

Fields

headId :: HeadId

decommitTxId :: TxIdType tx

utxoToDecommit :: UTxOType tx

DecommitFinalized

Fields

headId :: HeadId

distributedUTxO :: UTxOType tx

CommitRecorded

Fields

headId :: HeadId

utxoToCommit :: UTxOType tx

pendingDeposit :: TxIdType tx

deadline :: UTCTime

DepositActivated

Fields

headId :: HeadId

depositTxId :: TxIdType tx

deadline :: UTCTime

chainTime :: UTCTime

DepositExpired

Fields

headId :: HeadId

depositTxId :: TxIdType tx

deadline :: UTCTime

chainTime :: UTCTime

CommitApproved

Fields

headId :: HeadId

utxoToCommit :: UTxOType tx

CommitFinalized

Fields

headId :: HeadId

depositTxId :: TxIdType tx

CommitRecovered

Fields

headId :: HeadId

recoveredUTxO :: UTxOType tx

recoveredTxId :: TxIdType tx

SnapshotSideLoaded
Snapshot was side-loaded, and the included transactions can be considered final. The local state has been reset, meaning pending transactions were pruned. Any signing round has been discarded, and the snapshot leader has changed accordingly.

Fields

headId :: HeadId

snapshotNumber :: SnapshotNumber

EventLogRotated

Fields

checkpoint :: NodeState tx

NodeUnsynced

NodeSynced

data WithUTxO

Whether or not to include full UTxO in server outputs.

Constructors

WithUTxO

WithoutUTxO

data WithAddressedTx

Whether or not to filter transaction server outputs by given address.

Constructors

WithAddressedTx Text

WithoutAddressedTx

data ServerOutputConfig

Constructors

ServerOutputConfig

Fields

utxoInSnapshot :: WithUTxO

addressInTx :: WithAddressedTx

prepareServerOutput

Arguments

:: IsChainState tx

=> ServerOutputConfig
Decide on tx representation

-> TimedServerOutput tx
Server output

-> ByteString
Final output

Replaces the json encoded tx field with it's cbor representation.

NOTE: we deliberately pattern match on all ServerOutput constructors in handleTxOutput so that we don't forget to update this function if they change.

removeSnapshotUTxO :: ByteString -> ByteString

handleUtxoInclusion :: ServerOutputConfig -> (a -> a) -> a -> a

data HeadStatus

All possible Hydra states displayed in the API server outputs.

Constructors

Idle

Initializing

Open

Closed

FanoutPossible

data CommitInfo

All information needed to distinguish behavior of the commit endpoint.

Constructors

CannotCommit

NormalCommit HeadId

IncrementalCommit HeadId

data NetworkInfo

L2 Hydra network status information.

Constructors

NetworkInfo

Fields

networkConnected :: Bool

peersInfo :: Map Host Bool

getSnapshotUtxo :: Monoid (UTxOType tx) => HeadState tx -> Maybe (UTxOType tx)

Get latest confirmed snapshot UTxO from HeadState.

getSeenSnapshot :: HeadState tx -> SeenSnapshot tx

Get latest seen snapshot from HeadState.

getConfirmedSnapshot :: IsChainState tx => HeadState tx -> Maybe (ConfirmedSnapshot tx)

Get latest confirmed snapshot from HeadState.