Hydra.BehaviorSpec

Documentation

spec :: Spec

waitUntil :: (HasCallStack, MonadThrow m, MonadAsync m, MonadTimer m, MonadLabelledSTM m, IsChainState tx) => [TestHydraClient tx m] -> ServerOutput tx -> m ()

Wait for some output at some node(s) to be produced eventually. See waitUntilMatch for how long it waits.

waitUntilMatch :: (Show (ServerOutput tx), HasCallStack, MonadThrow m, MonadAsync m, MonadTimer m, MonadLabelledSTM m, Eq a, Show a, IsChainState tx) => [TestHydraClient tx m] -> (ServerOutput tx -> Maybe a) -> m a

Wait for a server output to match some predicate eventually. If a client message is received instead, this fails. This will not wait forever, but for a long time (1 month) to get a nice error location. Should not be an issue when used within shouldRunInSim, this was even 1000 years before - but we since we are having the protocol produce Tick events constantly this would be fully simulated to the end.

newtype NodeId

Constructors

Node Natural

data TestHydraClient tx m

A thin client layer around HydraNode to be interact with it through $sel:send:TestHydraClient, $sel:waitForNext:TestHydraClient, access all outputs and inject events through the test chain.

Constructors

TestHydraClient

Fields

send :: ClientInput tx -> m ()

waitForNext :: m (ServerOutput tx)

waitForNextMessage :: m (ClientMessage tx)

injectChainEvent :: ChainEvent tx -> m ()

serverOutputs :: m [ServerOutput tx]

queryState :: m (NodeState tx)

data SimulatedChainNetwork tx m

A simulated chain that just echoes PostChainTx as Observations of OnChainTx onto all connected nodes. It can also $sel:rollbackAndForward:SimulatedChainNetwork any number of these "transactions".

Constructors

SimulatedChainNetwork

Fields

connectNode :: DraftHydraNode tx m -> m (HydraNode tx m)

tickThread :: Async m ()

rollbackAndForward :: Natural -> m ()

simulateCommit :: HeadId -> Party -> UTxOType tx -> m ()

simulateDeposit :: HeadId -> UTxOType tx -> UTCTime -> m (TxIdType tx)

closeWithInitialSnapshot :: (Party, UTxOType tx) -> m ()

dummySimulatedChainNetwork :: SimulatedChainNetwork tx m

withSimulatedChainAndNetwork :: (MonadTime m, MonadDelay m, MonadAsync m, MonadThrow m, MonadLabelledSTM m) => (SimulatedChainNetwork SimpleTx m -> m a) -> m a

With-pattern wrapper around simulatedChainAndNetwork which does cancel the $sel:tickThread:SimulatedChainNetwork. Also, this will fix tx to SimpleTx so that it can pick an initial chain state to play back to our test nodes. NOTE: The simulated network has a block time of 20 (simulated) seconds.

simulatedChainAndNetwork :: forall m. (MonadTime m, MonadDelay m, MonadAsync m, MonadLabelledSTM m) => ChainStateType SimpleTx -> m (SimulatedChainNetwork SimpleTx m)

Creates a simulated chain and network to which HydraNodes can be connected to using $sel:connectNode:SimulatedChainNetwork. NOTE: The $sel:tickThread:SimulatedChainNetwork needs to be canceled after use. Use withSimulatedChainAndNetwork instead where possible.

handleChainEvent :: HydraNode tx m -> ChainEvent tx -> m ()

createMockNetwork :: MonadSTM m => DraftHydraNode tx m -> TVar m [HydraNode tx m] -> Network m (Message tx)

toOnChainTx :: IsTx tx => UTCTime -> PostChainTx tx -> OnChainTx tx

Derive an OnChainTx from PostChainTx to simulate a "perfect" chain. NOTE: This implementation announces hard-coded contestationDeadlines. Also, all heads will have the same OpenState and OpenState.

newDeadlineFarEnoughFromNow :: MonadTime m => m UTCTime

nothingHappensFor :: (MonadTimer m, MonadThrow m, IsChainState tx) => TestHydraClient tx m -> NominalDiffTime -> m ()

withHydraNode :: forall s a. SigningKey HydraKey -> [Party] -> SimulatedChainNetwork SimpleTx (IOSim s) -> (TestHydraClient SimpleTx (IOSim s) -> IOSim s a) -> IOSim s a

withHydraNode' :: DepositPeriod -> SigningKey HydraKey -> [Party] -> SimulatedChainNetwork SimpleTx (IOSim s) -> (TestHydraClient SimpleTx (IOSim s) -> IOSim s b) -> IOSim s b

createTestHydraClient :: MonadSTM m => TQueue m (ServerOutput tx) -> TQueue m (ClientMessage tx) -> TVar m [ServerOutput tx] -> HydraNode tx m -> TestHydraClient tx m

createHydraNode :: (IsChainState tx, MonadDelay m, MonadAsync m, MonadLabelledSTM m, MonadThrow m) => Tracer m (HydraNodeLog tx) -> Ledger tx -> ChainStateType tx -> SigningKey HydraKey -> [Party] -> TQueue m (ServerOutput tx) -> TQueue m (ClientMessage tx) -> TVar m [ServerOutput tx] -> SimulatedChainNetwork tx m -> ContestationPeriod -> DepositPeriod -> m (HydraNode tx m)

openHead :: SimulatedChainNetwork SimpleTx (IOSim s) -> TestHydraClient SimpleTx (IOSim s) -> IOSim s ()

openHead2 :: SimulatedChainNetwork SimpleTx (IOSim s) -> TestHydraClient SimpleTx (IOSim s) -> TestHydraClient SimpleTx (IOSim s) -> IOSim s ()

assertHeadIsClosed :: (HasCallStack, MonadThrow m) => ServerOutput tx -> m ()

assertHeadIsClosedWith :: (HasCallStack, MonadThrow m) => SnapshotNumber -> ServerOutput tx -> m ()

shortLabel :: SigningKey HydraKey -> String

Provide a quick and dirty to way to label stuff from a signing key

getHeadUTxO :: IsTx tx => HeadState tx -> Maybe (UTxOType tx)

Get the head UTxO from open HeadState.