Hydra.Cardano.Api.TxIn

Extras

mkTxIn :: Tx era -> Word -> TxIn

Create a TxIn (a.k.a UTXO) from a transaction and output index.

withWitness :: TxIn -> (TxIn, BuildTxWith BuildTx (Witness WitCtxTxIn Era))

Attach some verification-key witness to a TxIn.

txIns' :: Tx era -> [TxIn]

Access inputs of a transaction, as an ordered list.

txInputSet :: Tx era -> Set TxIn

Access inputs of a transaction, as an ordered set.

Type Conversions

fromLedgerTxIn :: TxIn -> TxIn

Convert a cardano-ledger TxIn into a cardano-api TxIn

toLedgerTxIn :: TxIn -> TxIn

Convert a cardano-api TxIn into a cardano-ledger TxIn

fromPlutusTxOutRef :: TxOutRef -> TxIn

Convert a plutus' TxOutRef into a cardano-api TxIn

toPlutusTxOutRef :: TxIn -> TxOutRef

Convert a cardano-api TxIn into a plutus TxOutRef.

Arbitrary values

genTxIn :: Gen TxIn

A more random generator than the 'Arbitrary TxIn' from cardano-ledger. NOTE: This is using the Cardano ledger's deserialization framework using the latest protocol version via maxBound.

Orphan instances

Arbitrary TxIn

Instance details

Methods

arbitrary :: Gen TxIn

shrink :: TxIn -> [TxIn]