Hydra.Chain

Description

Specifies the Head-Chain Interaction part of the protocol

Incoming and outgoing on-chain transactions are modelled respectively as OnChainTx and PostChainTx which are data type that abstracts away the details of the structure of the transaction.

Documentation

maxMainnetLovelace :: Coin

Hardcoded limit for commit tx on mainnet

maximumNumberOfParties :: Int

Hardcoded limit for maximum number of parties in a head protocol The value is obtained from calculating the costs of running the scripts and on-chan validators (see computeCollectComCost computeAbortCost). A too high enough number would be detected by property and acceptance tests.

data PostChainTx tx

Data type used to post transactions on chain. It holds everything to construct corresponding Head protocol transactions.

Constructors

InitTx

Fields

participants :: [OnChainId]

headParameters :: HeadParameters

AbortTx

Fields

utxo :: UTxOType tx

headSeed :: HeadSeed

CollectComTx

Fields

utxo :: UTxOType tx

headId :: HeadId

headParameters :: HeadParameters

IncrementTx

Fields

headId :: HeadId

headParameters :: HeadParameters

incrementingSnapshot :: ConfirmedSnapshot tx

depositTxId :: TxIdType tx

RecoverTx

Fields

headId :: HeadId

recoverTxId :: TxIdType tx

deadline :: ChainSlot

recoverUTxO :: UTxOType tx

DecrementTx

Fields

headId :: HeadId

headParameters :: HeadParameters

decrementingSnapshot :: ConfirmedSnapshot tx

CloseTx

Fields

headId :: HeadId

headParameters :: HeadParameters

openVersion :: SnapshotVersion

closingSnapshot :: ConfirmedSnapshot tx

ContestTx

Fields

headId :: HeadId

headParameters :: HeadParameters

openVersion :: SnapshotVersion

contestingSnapshot :: ConfirmedSnapshot tx

FanoutTx

Fields

utxo :: UTxOType tx

utxoToCommit :: Maybe (UTxOType tx)

utxoToDecommit :: Maybe (UTxOType tx)

headSeed :: HeadSeed

contestationDeadline :: UTCTime

data OnChainTx tx

Describes transactions as seen on chain. Holds as minimal information as possible to simplify observing the chain.

Constructors

OnInitTx

Fields

headId :: HeadId

headSeed :: HeadSeed

headParameters :: HeadParameters

participants :: [OnChainId]

OnCommitTx

Fields

headId :: HeadId

party :: Party

committed :: UTxOType tx

OnAbortTx

Fields

headId :: HeadId

OnCollectComTx

Fields

headId :: HeadId

OnDepositTx

Fields

headId :: HeadId

depositTxId :: TxIdType tx

deposited :: UTxOType tx

created :: UTCTime

deadline :: UTCTime

OnRecoverTx

Fields

headId :: HeadId

recoveredTxId :: TxIdType tx

recoveredUTxO :: UTxOType tx

OnIncrementTx

Fields

headId :: HeadId

newVersion :: SnapshotVersion

depositTxId :: TxIdType tx

OnDecrementTx

Fields

headId :: HeadId

newVersion :: SnapshotVersion

distributedUTxO :: UTxOType tx

OnCloseTx

Fields

headId :: HeadId

snapshotNumber :: SnapshotNumber

contestationDeadline :: UTCTime

OnContestTx

Fields

headId :: HeadId

snapshotNumber :: SnapshotNumber

contestationDeadline :: UTCTime

OnFanoutTx

Fields

headId :: HeadId

fanoutUTxO :: UTxOType tx

data PostTxError tx

Exceptions thrown by $sel:postTx:Chain.

Constructors

NoSeedInput

InvalidSeed

Fields

headSeed :: HeadSeed

InvalidHeadId

Fields

headId :: HeadId

CannotFindOwnInitial

Fields

knownUTxO :: UTxOType tx

UnsupportedLegacyOutput
Committing byron addresses is not supported.

Fields

byronAddress :: Address ByronAddr

InvalidStateToPost

Fields

txTried :: PostChainTx tx

chainState :: ChainStateType tx

NotEnoughFuel

Fields

failingTx :: tx

NoFuelUTXOFound

Fields

failingTx :: tx

ScriptFailedInWallet
Script execution failed when finalizing a transaction in the wallet. XXX: Ideally we want a cardano-api type with corresponding JSON instance here. But the wallet still uses ledger types and we don't want to copy the conversion from ledger TransactionScriptFailure to the cardano-api ScriptExecutionError type.

Fields

redeemerPtr :: Text

failureReason :: Text

failingTx :: tx

InternalWalletError
A generic error happened when finalizing a transaction in the wallet.

Fields

headUTxO :: UTxOType tx

reason :: Text

failingTx :: tx

FailedToPostTx
An error occurred when submitting a transaction to the cardano-node.

Fields

failureReason :: Text

failingTx :: tx

CommittedTooMuchADAForMainnet
User tried to commit more than maxMainnetLovelace hardcoded limit on mainnet we keep track of both the hardcoded limit and what the user originally tried to commit

Fields

userCommittedLovelace :: Coin

mainnetLimitLovelace :: Coin

FailedToDraftTxNotInitializing
We can only draft commit tx for the user when in Initializing state

FailedToConstructAbortTx

FailedToConstructCloseTx

FailedToConstructContestTx

FailedToConstructCollectTx

FailedToConstructDepositTx

Fields

failureReason :: Text

FailedToConstructRecoverTx

Fields

failureReason :: Text

FailedToConstructIncrementTx

Fields

failureReason :: Text

FailedToConstructDecrementTx

Fields

failureReason :: Text

FailedToConstructFanoutTx

DepositTooLow

Fields

providedValue :: Coin

minimumValue :: Coin

AmountTooLow

Fields

providedValue :: Coin

totalUTxOValue :: Coin

InvalidTokenRequest [(PolicyId, PolicyAssets)]

data ChainStateHistory tx

A non empty sequence of chain states that can be rolled back. This is expected to be constructed by using the smart constructor initHistory.

Constructors

UnsafeChainStateHistory

Fields

history :: NonEmpty (ChainStateType tx)

The sequence of known chain states, ordered from most recent to oldest. These contain notable state observed on-chain to be able to interact with Hydra heads.

lastKnown :: ChainPointType tx

The last known chain point, which may be used to continue observing the chain.

defaultChainState :: ChainStateType tx

The default chain state to fall back to when rolling back beyond known history.

currentState :: ChainStateHistory tx -> ChainStateType tx

pushNewState :: IsChainState tx => ChainStateType tx -> ChainStateHistory tx -> ChainStateHistory tx

Record a new chain state in history. Also ensures the $sel:lastKnown:UnsafeChainStateHistory point is updated accordingly.

setLastKnown :: ChainPointType tx -> ChainStateHistory tx -> ChainStateHistory tx

Update the last known chain point. Use pushNewState if you have a full 'ChainStateType tx'.

initHistory :: IsChainState tx => ChainStateType tx -> ChainStateHistory tx

rollbackHistory :: IsChainState tx => ChainSlot -> ChainStateHistory tx -> ChainStateHistory tx

prefixOf :: IsChainState tx => ChainStateHistory tx -> NonEmpty (ChainPointType tx)

Get the known prefix of all the ChainStateHistory.

data Chain tx m

Handle to interface with the main chain network

Constructors

Chain

Fields

postTx :: MonadThrow m => PostChainTx tx -> m ()

Construct and send a transaction to the main chain corresponding to the given PostChainTx description. This function is not expected to block, so it is only responsible for submitting, but it should validate the created transaction against a reasonable local view of the chain and throw an exception when invalid.

Does at least throw PostTxError.

draftCommitTx :: MonadThrow m => HeadId -> CommitBlueprintTx tx -> m (Either (PostTxError tx) tx)

Create a commit transaction using user provided utxos (zero or many) and a _blueprint_ transaction which spends these outputs. Errors are handled at the call site.

draftDepositTx :: MonadThrow m => HeadId -> PParams LedgerEra -> CommitBlueprintTx tx -> UTCTime -> Maybe AddressInEra -> m (Either (PostTxError tx) tx)

Create a deposit transaction using user provided utxos (zero or many) , _blueprint_ transaction which spends these outputs and a deadline for their inclusion into L2. Errors are handled at the call site.

submitTx :: MonadThrow m => tx -> m ()

Submit a cardano transaction.

Throws at least PostTxError.

XXX: While technically they could be any of 'PostTxError tx', only FailedToPostTx errors are expected here.

checkNonADAAssets :: ConfirmedSnapshot tx -> Either Value ()

data ChainEvent tx

Constructors

Observation
Indicates a head protocol transaction has been observed.

Fields

observedTx :: OnChainTx tx

newChainState :: ChainStateType tx

Rollback

Fields

chainTime :: UTCTime

rolledBackChainState :: ChainStateType tx

Tick
Indicate time has advanced on the chain. This is deliberately not a ChainStateType because state updates are only expected upon Observation orRollback.

NOTE: While the type does not guarantee that the UTCTime and the slot in ChainPointType tx are consistent the alternative would be provide the means to do the conversion. For Cardano, this would be a systemStart and eraHistory.. which is annoying and if it's kept in the chain layer, it would mean another round trip / state to keep there.

Fields

chainTime :: UTCTime

chainPoint :: ChainPointType tx

PostTxError
Event to re-ingest errors from $sel:postTx:Chain for further processing.

Fields

postChainTx :: PostChainTx tx

postTxError :: PostTxError tx

failingTx :: Maybe tx

type ChainCallback tx m = ChainEvent tx -> m ()

A callback indicating a 'ChainEvent tx' happened. Most importantly the Observation of a relevant Hydra transaction.

type ChainComponent tx m a = ChainCallback tx m -> (Chain tx m -> m a) -> m a

A type tying both posting and observing transactions into a single Component.