Hydra.Chain.Blockfrost.Client

Documentation

newtype APIBlockfrostError

Constructors

BlockfrostError BlockfrostException

data BlockfrostException

Constructors

TimeoutOnUTxO TxId

NoUTxOFound (Address ShelleyAddr)

FailedToDecodeAddress Text

ByronAddressNotSupported

FailedUTxOForHash Text

FailedEraHistory

AssetNameMissing

DeserialiseError Text

DecodeError Text

BlockfrostAPIError Text

data BlockfrostConversion

Constructors

BlockfrostConversion

Fields

a0 :: NonNegativeInterval

rho :: UnitInterval

tau :: UnitInterval

priceMemory :: NonNegativeInterval

priceSteps :: NonNegativeInterval

pvtMotionNoConfidence :: UnitInterval

pvtCommitteeNormal :: UnitInterval

pvtCommitteeNoConfidence :: UnitInterval

pvtHardForkInitiation :: UnitInterval

pvtPPSecurityGroup :: UnitInterval

dvtMotionNoConfidence :: UnitInterval

dvtCommitteeNormal :: UnitInterval

dvtCommitteeNoConfidence :: UnitInterval

dvtUpdateToConstitution :: UnitInterval

dvtHardForkInitiation :: UnitInterval

dvtPPNetworkGroup :: UnitInterval

dvtPPEconomicGroup :: UnitInterval

dvtPPTechnicalGroup :: UnitInterval

dvtPPGovGroup :: UnitInterval

dvtTreasuryWithdrawal :: UnitInterval

committeeMinSize :: Quantity

committeeMaxTermLength :: Quantity

govActionLifetime :: Quantity

govActionDeposit :: Coin

drepDeposit :: Integer

drepActivity :: Quantity

minFeeRefScriptCostPerByte :: NonNegativeInterval

queryGenesisParameters :: BlockfrostClientT IO Genesis

Query the Blockfrost API for Genesis

queryScriptRegistry :: BlockfrostOptions -> [TxId] -> BlockfrostClientT IO ScriptRegistry

Query for TxIns in the search for outputs containing all the reference scripts of the ScriptRegistry.

This is implemented by repeated querying until we have all necessary reference scripts as we do only know the transaction id, not the indices.

Can throw at least NewScriptRegistryException on failure.

queryEraHistory :: BlockfrostClientT IO EraHistory

queryProtocolParameters :: MonadIO m => BlockfrostClientT m (PParams LedgerEra)

queryStakePools :: BlockfrostClientT IO (Set PoolId)

querySystemStart :: BlockfrostClientT IO SystemStart

submitTransaction :: MonadIO m => Tx -> BlockfrostClientT m TxHash

awaitTransaction :: BlockfrostOptions -> Tx -> VerificationKey PaymentKey -> BlockfrostClientT IO UTxO

queryUTxOByTxIn :: BlockfrostOptions -> NetworkId -> [TxIn] -> BlockfrostClientT IO UTxO

Query the Blockfrost API to get the UTxO for TxIn and convert to cardano UTxO. FIXME: make blockfrost wait times configurable.

queryTip :: BlockfrostClientT IO ChainPoint

Query the Blockfrost API for Genesis and convert to cardano ChainPoint.

queryUTxO :: BlockfrostOptions -> NetworkId -> [Address ShelleyAddr] -> BlockfrostClientT IO UTxO

Query the Blockfrost API for address UTxO and convert to cardano UTxO. NOTE: We accept the address list here to be compatible with cardano-api but in fact this is a single address query always.

queryUTxOFor :: BlockfrostOptions -> VerificationKey PaymentKey -> BlockfrostClientT IO UTxO

runBlockfrostM :: (MonadIO m, MonadThrow m) => Project -> BlockfrostClientT IO a -> m a

toCardanoNetworkId :: Integer -> NetworkId

scriptTypeToPlutusVersion :: ScriptType -> Maybe Language

toCardanoUTxO :: NetworkId -> TxIn -> Address -> Maybe ScriptHash -> Maybe DatumHash -> [Amount] -> Maybe InlineDatum -> BlockfrostClientT IO UTxO

toCardanoValue :: [Amount] -> BlockfrostClientT IO Value

queryScript :: Text -> BlockfrostClientT IO (Maybe PlutusScript)

toCardanoTxOut :: NetworkId -> Text -> Value -> Maybe Text -> Maybe Text -> Maybe PlutusScript -> BlockfrostClientT IO (TxOut ctx)

toCardanoPoolId :: PoolId -> Hash StakePoolKey

toCardanoTxIn :: Text -> Integer -> TxIn

toCardanoAddress :: Text -> Maybe AddressInEra

toCardanoPolicyIdAndAssetName :: Text -> BlockfrostClientT IO (PolicyId, AssetName)

toCardanoGenesisParameters :: Genesis -> GenesisParameters ShelleyEra

awaitUTxO

Arguments

:: NetworkId
Network id

-> [Address ShelleyAddr]
Address we are interested in

-> TxId
Last transaction ID to await

-> BlockfrostOptions

-> BlockfrostClientT IO UTxO

Await for specific UTxO at address - the one that is produced by the given TxId.

data Block

Constructors

Block

Fields

_blockTime :: POSIXTime

_blockHeight :: Maybe Integer

_blockHash :: BlockHash

_blockSlot :: Maybe Slot

_blockEpoch :: Maybe Epoch

_blockEpochSlot :: Maybe Integer

_blockSlotLeader :: Text

_blockSize :: Integer

_blockTxCount :: Integer

_blockOutput :: Maybe Lovelaces

_blockFees :: Maybe Lovelaces

_blockBlockVrf :: Maybe Text

_blockOpCert :: Maybe Text

_blockOpCertCounter :: Maybe Quantity

_blockPreviousBlock :: Maybe BlockHash

_blockNextBlock :: Maybe BlockHash

_blockConfirmations :: Integer

newtype Slot

Constructors

Slot Integer

newtype BlockHash

Constructors

BlockHash Text

data BlockfrostClientT (m :: Type -> Type) a

data Genesis

Constructors

Genesis

Fields

_genesisActiveSlotsCoefficient :: Rational

_genesisUpdateQuorum :: Integer

_genesisMaxLovelaceSupply :: Lovelaces

_genesisNetworkMagic :: Integer

_genesisEpochLength :: Integer

_genesisSystemStart :: POSIXTime

_genesisSlotsPerKesPeriod :: Integer

_genesisSlotLength :: Integer

_genesisMaxKesEvolutions :: Integer

_genesisSecurityParam :: Integer

data Project

newtype TransactionCBOR

Constructors

TransactionCBOR

Fields

_transactionCBORCbor :: Text

newtype TxHashCBOR

Constructors

TxHashCBOR

Fields

getTxHashCBOR :: (TxHash, TransactionCBOR)

tryError :: MonadError e m => m a -> m (Either e a)

MonadError analogue to the try function.

allPages :: Monad m => (Paged -> m [a]) -> m [a]

def :: Default a => a

getBlock :: MonadBlockfrost m => Either Integer BlockHash -> m Block

getBlockTxsCBOR' :: MonadBlockfrost m => Either Integer BlockHash -> Paged -> SortOrder -> m [TxHashCBOR]

getLedgerGenesis :: MonadBlockfrost m => m Genesis

listPools :: MonadBlockfrost m => m [PoolId]

projectFromFile :: FilePath -> IO Project

runBlockfrost :: Project -> BlockfrostClientT IO a -> IO (Either BlockfrostError a)

unBlockHash :: BlockHash -> Text

unSlot :: Slot -> Integer