Hydra.Chain.CardanoClient

Description

A basic cardano-node client that can talk to a local cardano-node.

The idea of this module is to provide a Haskell interface on top of cardano-cli's API, using cardano-api types.

Documentation

data QueryException

Constructors

QueryAcquireException AcquiringFailure

QueryEraMismatchException EraMismatch

QueryUnsupportedNtcVersionException UnsupportedNtcVersionError

QueryProtocolParamsConversionException ProtocolParametersConversionError

QueryProtocolParamsEraNotSupported AnyCardanoEra

QueryProtocolParamsEncodingFailureOnEra AnyCardanoEra Text

QueryEraNotInCardanoModeFailure AnyCardanoEra

QueryNotShelleyBasedEraException AnyCardanoEra

data SubmitTransactionException

Exceptions that can occur during a transaction submission.

In principle, we can only encounter an EraMismatch at era boundaries, when we try to submit a "next era" transaction as a "current era" transaction, or vice-versa. Similarly, TxValidationError shouldn't occur given that the transaction was safely constructed through buildTransaction.

Constructors

SubmitEraMismatch EraMismatch

SubmitTxValidationError TxValidationErrorInCardanoMode

CardanoClient handle

data CardanoClient era

Handle interface for abstract querying of a cardano node.

Constructors

CardanoClient

Fields

queryUTxOByAddress :: [Address ShelleyAddr] -> IO (UTxO era)

networkId :: NetworkId

Local state query

data QueryPoint

Describes whether to query at the tip or at a specific point.

Constructors

QueryTip

QueryAt ChainPoint

CardanoNode Monad

newtype CardanoNodeT eon era m a

Monad transformer for interacting with a local Cardano node. Carries both LocalNodeConnectInfo and an era witness so individual query and submission functions need not accept them as explicit arguments.

Constructors

CardanoNodeT

Fields

fromCardanoNodeT :: ReaderT LocalNodeConnectInfo (EraT eon era m) a

type CardanoNode eon era a = CardanoNodeT eon era IO a

Specialise CardanoNodeT to IO.

EraT Monad Transformer

newtype EraT eon era m a

Monad transformer that carries an era witness. The eon parameter is a higher-kinded era-witness type (e.g. ShelleyBasedEra, BabbageEraOnwards, ConwayEraOnwards) and era is the concrete era (e.g. CardanoEra).

Constructors

EraT

Fields

fromEraT :: ReaderT (eon era) m a

runEraT :: eon era -> EraT eon era m a -> m a

askEra :: Monad m => CardanoNodeT eon era m (eon era)

runCardanoNodeT :: LocalNodeConnectInfo -> eon era -> CardanoNodeT eon era m a -> m a

Run a CardanoNodeIO action given a connection and era witness.

runCardanoNode :: LocalNodeConnectInfo -> eon era -> CardanoNode eon era a -> IO a

Tx Construction / Submission

submitTransaction :: MonadIO m => Tx -> CardanoNodeT ShelleyBasedEra Era m ()

Submit a (signed) transaction to the node.

Throws SubmitTransactionException if submission fails.

awaitTransaction :: MonadIO m => Tx -> CardanoNodeT ShelleyBasedEra era m (UTxO era)

Await until the given transaction is visible on-chain. Returns the UTxO set produced by that transaction.

Note that this function loops forever; hence, one probably wants to couple it with a surrounding timeout.

queryTip :: MonadIO m => CardanoNodeT eon era m ChainPoint

Query the latest chain point aka "the tip".

querySystemStart :: MonadIO m => QueryPoint -> CardanoNodeT eon era m SystemStart

Query the system start parameter at given point.

Throws at least QueryException if query fails.

queryEraHistory :: MonadIO m => QueryPoint -> CardanoNodeT eon era m EraHistory

Query the era history at given point.

Throws at least QueryException if query fails.

queryEpochNo :: MonadIO m => QueryPoint -> CardanoNodeT ShelleyBasedEra era m EpochNo

Query the current epoch number.

Throws at least QueryException if query fails.

queryProtocolParameters :: MonadIO m => QueryPoint -> CardanoNodeT ShelleyBasedEra era m (PParams (ShelleyLedgerEra era))

Query the protocol parameters at given point.

Throws at least QueryException if query fails.

queryGenesisParameters :: MonadIO m => QueryPoint -> CardanoNodeT ShelleyBasedEra era m (GenesisParameters ShelleyEra)

Query GenesisParameters at a given point.

Throws at least QueryException if query fails.

queryUTxO :: MonadIO m => QueryPoint -> [Address ShelleyAddr] -> CardanoNodeT ShelleyBasedEra era m (UTxO era)

Query UTxO for all given addresses at given point.

Throws at least QueryException if query fails.

queryUTxOByTxIn :: MonadIO m => QueryPoint -> [TxIn] -> CardanoNodeT ShelleyBasedEra era m (UTxO era)

Query UTxO for given tx inputs at given point.

Throws at least QueryException if query fails.

queryUTxOWhole :: MonadIO m => QueryPoint -> CardanoNodeT ShelleyBasedEra era m (UTxO era)

Query the whole UTxO from node at given point. Useful for debugging, but should obviously not be used in production code.

Throws at least QueryException if query fails.

queryUTxOFor :: MonadIO m => QueryPoint -> VerificationKey PaymentKey -> CardanoNodeT ShelleyBasedEra era m (UTxO era)

Query UTxO for the address of given verification key at point.

Throws at least QueryException if query fails.

queryStakePools :: MonadIO m => QueryPoint -> CardanoNodeT ShelleyBasedEra era m (Set PoolId)

Query the current set of registered stake pools.

Throws at least QueryException if query fails.

Helpers

throwOnEraMismatch :: MonadThrow m => Either EraMismatch a -> m a

throwOnUnsupportedNtcVersion :: MonadThrow m => Either UnsupportedNtcVersionError a -> m a

liftQueryResult :: MonadThrow m => Either UnsupportedNtcVersionError (Either EraMismatch a) -> m a

Lift a query result that may have NTC version or era mismatch errors into the monad. This combines the common pattern of checking both error types.

localNodeConnectInfo :: NetworkId -> SocketPath -> LocalNodeConnectInfo

cardanoModeParams :: ConsensusModeParams

runQuery :: MonadIO m => QueryPoint -> LocalStateQueryExpr BlockInMode ChainPoint QueryInMode () IO a -> CardanoNodeT eon era m a

Execute a local-state query at the given point, using the connectInfo from the reader environment. Throws QueryAcquireException on failure.

queryCurrentEraExpr :: LocalStateQueryExpr b p QueryInMode r IO AnyCardanoEra

Query the current era from the node without needing an era witness.

runNodeIOWithCurrentEra :: NetworkId -> SocketPath -> (forall era. ShelleyBasedEra era -> CardanoNode ShelleyBasedEra era a) -> IO a

Run a CardanoNode action after dynamically discovering the node's current era. This is useful when connecting to a node whose era is not known in advance.

Throws QueryNotShelleyBasedEraException if the node is in Byron era.