Hydra.Chain.Direct.Contract.CollectCom

Documentation

healthyCollectComTx :: (Tx, UTxO)

healthyCommits :: Map TxIn HealthyCommit

healthyCommittedUTxO :: [UTxO]

healthyContestationPeriod :: ContestationPeriod

healthyHeadInput :: TxIn

healthyHeadResolvedInput :: TxOut CtxUTxO

healthyCollectComInitialDatum :: State

healthyOnChainParties :: [Party]

healthyParties :: [Party]

genCommittableTxOut :: Gen (TxIn, TxOut CtxUTxO)

data HealthyCommit

Constructors

HealthyCommit

Fields

cardanoKey :: VerificationKey PaymentKey

txOut :: TxOut CtxUTxO

scriptData :: HashableScriptData

healthyCommitOutput :: Party -> UTxO -> (TxIn, HealthyCommit)

data CollectComMutation

Constructors

NotContinueContract
Ensures collectCom does not allow any output address but νHead.

ExtractSomeValue
Needs to prevent that not all value is collected into the head output.

MutateOpenUTxOHash

MutateCommitToInitial
Ensures collectCom cannot collect from an initial UTxO.

MutateNumberOfParties
Every party should have commited and been taken into account for the collectCom transaction to be valid. Here we increase the number of parties in input and output but keep the commits unchanged. This simulates the situation where one participant would not have commited already or whose commit would have been ignored by the collectCom transaction.

MutateHeadId

MutateRequiredSigner

MutateTokenMintingOrBurning
Minting or burning of tokens should not be possible in collectCom.

genCollectComMutation :: (Tx, UTxO) -> Gen SomeMutation