Hydra.Chain.Direct.Contract.Commit

Description

Mutation-based script validator tests for the commit transaction where a healthyCommitTx gets mutated by an arbitrary CommitMutation.

Documentation

healthyCommitTx :: (Tx, UTxO)

commitVerificationKey :: VerificationKey PaymentKey

healthyIntialTxIn :: TxIn

healthyInitialTxOut :: TxOut CtxTx

healthyCommittedUTxO :: UTxO

data CommitMutation

Constructors

NonContinuousHeadId
The headId in the output datum must match the one from the input datum.

MutateCommitOutputValue
Invalidates the transaction by changing the committed output value.

MutateCommittedValue
Invalidates the transaction by changing the value of the committed utxo on the input side of the transaction.

MutateCommittedAddress
Ensures the datum recording the commit is consistent with the UTxO being committed.

RecordAllCommittedUTxO
Ensures a commit cannot be left out when "declared" in the commit transaction output datum.

MutateRequiredSigner
Ensures commit is authenticated by a Head party by changing the signer used on the transaction to be the one in the PT.

UsePTFromDifferentHead
Change the head policy id to simulate commit using a PT and signer from a different head. The signer shows a correct signature but from a different head. This will cause the signer to not be present in the participation tokens.

MutateTokenMintingOrBurning
Minting or burning of the tokens should not be possible in commit.

genCommitMutation :: (Tx, UTxO) -> Gen SomeMutation