Hydra.Chain.Direct.Contract.Gen

Description

Generators used in mutation testing framework

Party / key utilities

genForParty :: Gen a -> Party -> a

Generate some a given the Party as a seed. NOTE: While this is useful to generate party-specific values, it DOES depend on the generator used. For example, `genForParty genVerificationKey` and `genForParty (fst $ genKeyPair)` do not yield the same verification keys!

genBytes :: Gen ByteString

genHash :: Gen ByteString

genMintedOrBurnedValue :: Gen Value

Generates value such that: - alters between policy id we use in test fixtures with a random one. - mixing arbitrary token names with hydraHeadV1 - excluding 0 for quantity to mimic minting/burning