Hydra.Chain.Direct.Handlers

Description

Provide infrastructure-independent "handlers" for posting transactions and following the chain.

This module encapsulates the transformation logic between cardano transactions and HydraNode abstractions PostChainTx and OnChainTx, and maintenance of on-chain relevant state.

Documentation

data LocalChainState m tx

Handle of a mutable local chain state that is kept in the direct chain layer.

Constructors

LocalChainState

Fields

getLatest :: STM m (ChainStateType tx)

pushNew :: ChainStateType tx -> STM m ()

rollback :: ChainSlot -> STM m (ChainStateType tx)

history :: STM m (ChainStateHistory tx)

newLocalChainState :: forall m tx. (IsChainState tx, MonadLabelledSTM m) => ChainStateHistory tx -> m (LocalChainState m tx)

Initialize a new local chain state from a given chain state history.

Posting Transactions

type SubmitTx m = Tx -> m ()

A callback used to actually submit a transaction to the chain.

type GetTimeHandle m = m TimeHandle

A way to acquire a TimeHandle

mkChain

Arguments

:: (MonadSTM m, MonadThrow (STM m))

=> Tracer m CardanoChainLog

-> GetTimeHandle m
Means to acquire a new TimeHandle.

-> TinyWallet m

-> ChainContext

-> LocalChainState m Tx

-> SubmitTx m

-> Chain Tx m

Create a Chain component for posting "real" cardano transactions.

This component does not actually interact with a cardano-node, but creates cardano transactions from PostChainTx transactions emitted by a HydraNode, balancing and signing them using given TinyWallet, before handing it off to the given SubmitTx callback. There is also a draftTx option for drafting a commit tx on behalf of the user using their selected utxo.

NOTE: Given the constraints on m this function should work within IOSim and does not require any actual IO to happen which makes it highly suitable for simulations and testing.

rejectLowDeposits :: PParams LedgerEra -> UTxO -> Either (PostTxError Tx) ()

finalizeTx :: MonadThrow m => TinyWallet m -> ChainContext -> UTxO -> UTxO -> Tx -> m Tx

Balance and sign the given partial transaction.

Following the Chain

data ChainSyncHandler m

A handler that takes care of following the chain.

Constructors

ChainSyncHandler

Fields

onRollForward :: BlockHeader -> [Tx] -> m ()

onRollBackward :: ChainPoint -> m ()

data TimeConversionException

Conversion of a slot number to a time failed. This can be usually be considered an internal error and may be happening because the used era history is too old.

Constructors

TimeConversionException

Fields

slotNo :: SlotNo

reason :: Text

chainSyncHandler

Arguments

:: forall m. (MonadSTM m, MonadThrow m)

=> Tracer m CardanoChainLog
Tracer for logging

-> ChainCallback Tx m

-> GetTimeHandle m
Means to acquire a new TimeHandle.

-> ChainContext
Contextual information about our chain connection.

-> LocalChainState m Tx

-> ChainSyncHandler m
A chain-sync handler to use in a local-chain-sync client.

Creates a ChainSyncHandler that can notify the given callback of events happening on-chain.

This forms the other half of a ChainComponent along with mkChain but is decoupled from actual interactions with the chain.

A TimeHandle is needed to do `SlotNo -> POSIXTime` conversions for Tick events.

Throws TimeConversionException when a received block's SlotNo cannot be converted to a UTCTime with the given TimeHandle.

convertObservation :: TimeHandle -> HeadObservation -> Maybe (OnChainTx Tx)

prepareTxToPost

Arguments

:: forall m. (MonadSTM m, MonadThrow (STM m))

=> TimeHandle

-> TinyWallet m

-> ChainContext

-> UTxOType Tx
Spendable UTxO

-> PostChainTx Tx

-> STM m Tx

maxGraceTime :: NominalDiffTime

Maximum delay we put on the upper bound of transactions to fit into a block. NOTE: This is highly depending on the network. If the security parameter and epoch length result in a short horizon, this is problematic.

data CardanoChainLog

Constructors

ToPost

Fields

toPost :: PostChainTx Tx

PostingTx

Fields

txId :: TxId

PostedTx

Fields

txId :: TxId

PostingFailed

Fields

tx :: Tx

postTxError :: PostTxError Tx

RolledForward

Fields

point :: ChainPoint

receivedTxIds :: [TxId]

RolledBackward

Fields

point :: ChainPoint

Wallet TinyWalletLog