Hydra.Chain.Direct.State

Description

Contains the a stateful interface to transaction construction and observation.

It defines the 'ChainStateType tx' to be used in the Direct layer and it's constituents.

Documentation

class HasKnownUTxO a where

A class for accessing the known UTxO set in a type. This is useful to get all the relevant UTxO for resolving transaction inputs.

Methods

getKnownUTxO :: a -> UTxO

States & transitions

data ChainStateAt

The chain state used by the Hydra.Chain.Direct implementation. It records the actual ChainState paired with a ChainSlot (used to know up to which point to rewind on rollbacks). XXX: could move this into IsChainState and use UTxOType tx instead of ChainStateType tx

Constructors

ChainStateAt

Fields

spendableUTxO :: UTxO

recordedAt :: Maybe ChainPoint

chainSlotFromPoint :: ChainPoint -> ChainSlot

Get a generic ChainSlot from a Cardano ChainPoint. Slot 0 is used for the genesis point.

data ChainTransition

A definition of all transitions between ChainStates. Enumerable and bounded to be used as labels for checking coverage.

Constructors

Init

Abort

Commit

Collect

Increment

Decrement

Close

Contest

Fanout

data ChainState

An enumeration of all possible on-chain states of a Hydra Head, where each case stores the relevant information to construct & observe transactions to other states.

Constructors

Idle
The idle state does not contain any head-specific information and exists to be used as a starting and terminal state.

Initial InitialState

Open OpenState

Closed ClosedState

initialChainState :: ChainStateType Tx

Defines the starting state of the direct chain layer.

data ChainContext

Read-only chain-specific data. This is different to HydraContext as it only contains data known to single peer.

Constructors

ChainContext

Fields

networkId :: NetworkId

ownVerificationKey :: VerificationKey PaymentKey

ownParty :: Party

scriptRegistry :: ScriptRegistry

data InitialThreadOutput

Representation of the Head output after an Init transaction.

Constructors

InitialThreadOutput

Fields

initialThreadUTxO :: (TxIn, TxOut CtxUTxO)

initialContestationPeriod :: ContestationPeriod

initialParties :: [Party]

data InitialState

Constructors

InitialState

Fields

initialThreadOutput :: InitialThreadOutput

initialInitials :: [(TxIn, TxOut CtxUTxO)]

initialCommits :: [(TxIn, TxOut CtxUTxO)]

headId :: HeadId

seedTxIn :: TxIn

data OpenState

Constructors

OpenState

Fields

openUTxO :: UTxO

headId :: HeadId

seedTxIn :: TxIn

openUtxoHash :: UTxOHash

data ClosedState

Constructors

ClosedState

Fields

closedUTxO :: UTxO

headId :: HeadId

seedTxIn :: TxIn

contestationDeadline :: UTCTime

Constructing transactions

initialize

Arguments

:: ChainContext

-> TxIn
Seed input.

-> [OnChainId]
Verification key hashes of all participants.

-> HeadParameters

-> Tx

Construct an init transaction given some general ChainContext, the HeadParameters and a seed TxIn which will be spent.

commit

Arguments

:: ChainContext

-> HeadId

-> UTxO
Spendable UTxO

-> UTxO
UTxO to commit. All outputs are assumed to be owned by public keys

-> Either (PostTxError Tx) Tx

Construct a commit transaction based on known, spendable UTxO and some arbitrary UTxOs to commit. This does look for "our initial output" to spend and check the given UTxO to be compatible. Hence, this function does fail if already committed or if the head is not initializing.

NOTE: This version of commit does only commit outputs which are held by payment keys. For a variant which supports committing scripts, see commit'.

commit'

Arguments

:: ChainContext

-> HeadId

-> UTxO
Spendable UTxO

-> CommitBlueprintTx Tx

-> Either (PostTxError Tx) Tx

Construct a commit transaction based on known, spendable UTxO and some user UTxO inputs to commit. This does look for "our initial output" to spend and check the given UTxO to be compatible. Hence, this function does fail if already committed or if the head is not initializing.

NOTE: A simpler variant only supporting pubkey outputs is commit.

rejectByronAddress :: UTxO -> Either (PostTxError Tx) ()

rejectMoreThanMainnetLimit :: NetworkId -> UTxO -> Either (PostTxError Tx) ()

abort

Arguments

:: ChainContext

-> TxIn
Seed TxIn

-> UTxO
Spendable UTxO containing head, initial and commit outputs

-> UTxO
Committed UTxOs to reimburse.

-> Either AbortTxError Tx

Construct a abort transaction based on known, spendable UTxO. This function looks for head, initial and commit outputs to spend and it will fail if we can't find the head output.

data CollectTxError

Constructors

InvalidHeadIdInCollect

Fields

headId :: HeadId

CannotFindHeadOutputToCollect

collect

Arguments

:: ChainContext

-> HeadId

-> HeadParameters

-> UTxO
UTxO to be used to collect. Should match whatever is recorded in the commit inputs.

-> UTxO
Spendable UTxO containing head, initial and commit outputs

-> Either CollectTxError Tx

Construct a collect transaction based on known, spendable UTxO. This function looks for head output and commit outputs to spend and it will fail if we can't find the head output.

data IncrementTxError

Constructors

InvalidHeadIdInIncrement

Fields

headId :: HeadId

CannotFindHeadOutputInIncrement

CannotFindDepositOutputInIncrement

Fields

depositTxId :: TxId

SnapshotMissingIncrementUTxO

SnapshotIncrementUTxOIsNull

increment

Arguments

:: ChainContext

-> UTxO
Spendable UTxO containing head and deposit outputs

-> HeadId

-> HeadParameters

-> ConfirmedSnapshot Tx
Snapshot to increment with.

-> TxId
Deposited TxId

-> SlotNo
Valid until, must be before deadline.

-> Either IncrementTxError Tx

Construct a increment transaction spending the head and deposit outputs in given UTxO, and producing single head output for pending $sel:utxoToCommit:Snapshot of given Snapshot.

data DecrementTxError

Possible errors when trying to construct decrement tx

Constructors

InvalidHeadIdInDecrement

Fields

headId :: HeadId

CannotFindHeadOutputInDecrement

DecrementValueNegative

SnapshotDecrementUTxOIsNull

decrement

Arguments

:: ChainContext

-> UTxO
Spendable UTxO containing head, initial and commit outputs

-> HeadId

-> HeadParameters

-> ConfirmedSnapshot Tx
Snapshot to decrement with.

-> Either DecrementTxError Tx

Construct a decrement transaction spending the head output in given UTxO, and producing outputs for all pending $sel:utxoToDecommit:Snapshot of given Snapshot.

data CloseTxError

Constructors

InvalidHeadIdInClose

Fields

headId :: HeadId

CannotFindHeadOutputToClose

BothCommitAndDecommitInClose

data RecoverTxError

Constructors

InvalidHeadIdInRecover

Fields

headId :: HeadId

CannotFindDepositOutputToRecover

Fields

depositTxId :: TxId

CannotFindDepositedOutputToRecover

Fields

depositedTxId :: TxId

recover

Arguments

:: ChainContext

-> HeadId

-> TxId
Deposit TxId

-> UTxO
Spendable UTxO

-> SlotNo

-> Either RecoverTxError Tx

Construct a recover transaction spending the deposit output and producing outputs the user initially deposited.

close

Arguments

:: ChainContext

-> UTxO
Spendable UTxO containing head, initial and commit outputs

-> HeadId
Head id to close.

-> HeadParameters
Parameters of the head to close.

-> SnapshotVersion
Last known version of the open head. NOTE: We deliberately require a SnapshotVersion to be passed in, even though it could be extracted from the open head output in the spendable UTxO, to stay consistent with the way parameters are handled.

-> ConfirmedSnapshot Tx
Snapshot to close with.

-> SlotNo
Tx validity lower bound

-> PointInTime
Tx validity upper bound

-> Either CloseTxError Tx

Construct a close transaction spending the head output in given UTxO, head parameters, and a confirmed snapshot. NOTE: Lower and upper bound slot difference should not exceed contestation period.

data ContestTxError

Constructors

InvalidHeadIdInContest

Fields

headId :: HeadId

CannotFindHeadOutputToContest

MissingHeadDatumInContest

MissingHeadRedeemerInContest

WrongDatumInContest

FailedToConvertFromScriptDataInContest

BothCommitAndDecommitInContest

contest

Arguments

:: ChainContext

-> UTxO
Spendable UTxO containing head, initial and commit outputs

-> HeadId

-> ContestationPeriod

-> SnapshotVersion
Last known version of the open head. NOTE: We deliberately require a SnapshotVersion to be passed in, even though it could be extracted from the open head output in the spendable UTxO, to stay consistent with the way parameters are handled.

-> ConfirmedSnapshot Tx
Snapshot to contest with.

-> PointInTime
Current slot and posix time to be used as the contestation time.

-> Either ContestTxError Tx

Construct a contest transaction based on the ClosedState and a confirmed snapshot. The given PointInTime will be used as an upper validity bound and needs to be before the deadline.

data FanoutTxError

Constructors

CannotFindHeadOutputToFanout

MissingHeadDatumInFanout

WrongDatumInFanout

FailedToConvertFromScriptDataInFanout

BothCommitAndDecommitInFanout

fanout

Arguments

:: ChainContext

-> UTxO
Spendable UTxO containing head, initial and commit outputs

-> TxIn
Seed TxIn

-> UTxO
Snapshot UTxO to fanout

-> Maybe UTxO
Snapshot UTxO to commit to fanout

-> Maybe UTxO
Snapshot UTxO to decommit to fanout

-> SlotNo
Contestation deadline as SlotNo, used to set lower tx validity bound.

-> Either FanoutTxError Tx

Construct a fanout transaction based on the ClosedState and off-chain agreed UTxO set to fan out.

Helpers

utxoOfThisHead :: PolicyId -> UTxO -> UTxO

Observing Transitions

IdleState transitions

observeInit :: ChainContext -> [VerificationKey PaymentKey] -> Tx -> Either NotAnInitReason (OnChainTx Tx, InitialState)

Observe an init transition using a InitialState and observeInitTx.

InitialState transitions

observeCommit :: ChainContext -> InitialState -> Tx -> Maybe (OnChainTx Tx, InitialState)

Observe an commit transition using a InitialState and observeCommitTx. NOTE: This function is a bit fragile as it assumes commit output on first output while the underlying observeCommitTx could deal with commit outputs at any index. Only use this function in tests and benchmarks.

observeCollect :: InitialState -> Tx -> Maybe (OnChainTx Tx, OpenState)

Observe an collect transition using a InitialState and observeCollectComTx. This function checks the head id and ignores if not relevant.

OpenState transitions

observeClose :: OpenState -> Tx -> Maybe (OnChainTx Tx, ClosedState)

Observe a close transition using a OpenState and observeCloseTx. This function checks the head id and ignores if not relevant.

Generators

Warning zone

data HydraContext

Define some global context from which generators can pick values for generation. This allows to write fairly independent generators which however still make sense with one another within the context of a head.

For example, one can generate a head's _party_ from that global list, whereas other functions may rely on all parties and thus, we need both generation to be coherent.

Do not use this in production code, but only for generating test data.

Constructors

HydraContext

Fields

ctxVerificationKeys :: [VerificationKey PaymentKey]

ctxHydraSigningKeys :: [SigningKey HydraKey]

ctxNetworkId :: NetworkId

ctxContestationPeriod :: ContestationPeriod

ctxScriptRegistry :: ScriptRegistry

ctxParties :: HydraContext -> [Party]

ctxParticipants :: HydraContext -> [OnChainId]

ctxHeadParameters :: HydraContext -> HeadParameters

Danger zone

unsafeCommit

Arguments

:: HasCallStack

=> ChainContext

-> HeadId

-> UTxO
Spendable UTxO

-> UTxO
UTxO to commit. All outputs are assumed to be owned by public keys.

-> Tx

unsafeAbort

Arguments

:: HasCallStack

=> ChainContext

-> TxIn
Seed TxIn

-> UTxO
Spendable UTxO containing head, initial and commit outputs

-> UTxO
Committed UTxOs to reimburse.

-> Tx

unsafeIncrement

Arguments

:: HasCallStack

=> ChainContext

-> UTxO
Spendable UTxO

-> HeadId

-> HeadParameters

-> ConfirmedSnapshot Tx

-> TxId

-> SlotNo

-> Tx

unsafeDecrement

Arguments

:: HasCallStack

=> ChainContext

-> UTxO
Spendable UTxO

-> HeadId

-> HeadParameters

-> ConfirmedSnapshot Tx

-> Tx

unsafeClose

Arguments

:: HasCallStack

=> ChainContext

-> UTxO
Spendable UTxO containing head, initial and commit outputs

-> HeadId

-> HeadParameters

-> SnapshotVersion

-> ConfirmedSnapshot Tx

-> SlotNo

-> PointInTime

-> Tx

Unsafe version of close that throws an error if the transaction fails to build.

unsafeCollect

Arguments

:: ChainContext

-> HeadId

-> HeadParameters

-> UTxO
UTxO to be used to collect. Should match whatever is recorded in the commit inputs.

-> UTxO
Spendable UTxO containing head, initial and commit outputs

-> Tx

unsafeContest

Arguments

:: HasCallStack

=> ChainContext

-> UTxO
Spendable UTxO containing head, initial and commit outputs

-> HeadId

-> ContestationPeriod

-> SnapshotVersion

-> ConfirmedSnapshot Tx

-> PointInTime

-> Tx

Unsafe version of contest that throws an error if the transaction fails to build.

unsafeFanout

Arguments

:: HasCallStack

=> ChainContext

-> UTxO
Spendable UTxO containing head, initial and commit outputs

-> TxIn
Seed TxIn

-> UTxO
Snapshot UTxO to fanout

-> Maybe UTxO
Snapshot commit UTxO to fanout

-> Maybe UTxO
Snapshot decommit UTxO to fanout

-> SlotNo
Contestation deadline as SlotNo, used to set lower tx validity bound.

-> Tx

unsafeObserveInit :: HasCallStack => ChainContext -> [VerificationKey PaymentKey] -> Tx -> InitialState

unsafeObserveInitAndCommits :: HasCallStack => ChainContext -> [VerificationKey PaymentKey] -> Tx -> [Tx] -> ([UTxO], InitialState)

Orphan instances

IsChainState Tx

Instance details

Associated Types

type ChainPointType Tx = (p :: Type)

type ChainStateType Tx = (c :: Type)

Methods

chainStatePoint :: ChainStateType Tx -> ChainPointType Tx

chainPointSlot :: ChainPointType Tx -> ChainSlot