Hydra.Chain.Direct.StateSpec

Documentation

spec :: Spec

genInitTxMutation :: TxIn -> Tx -> Gen (Mutation, String, NotAnInitReason)

genCommitTxMutation :: UTxO -> Tx -> Gen Mutation

genAdaOnlyUTxOOnMainnetWithAmountBiggerThanOutLimit :: Gen UTxO

Properties

prop_observeAnyTx :: Property

Given any Head protocol state and the transaction corresponding a protocol transition we should be able to observe this transition correctly even in presence of other valid Hydra Head protocol states in the used lookup utxo.

prop_splitUTxO :: UTxO -> Property

Given a UTxO with more than one entry, we can split it into two non-empty UTxO.

prop_canCloseFanoutEveryCollect :: Property

prop_incrementObservesCorrectUTxO :: Property

propBelowSizeLimit :: Natural -> ((UTxO -> Tx -> Property) -> Property) -> SpecWith ()

propIsValid :: ((UTxO -> Tx -> Property) -> Property) -> SpecWith ()

Generators

forAllInit :: Testable property => (UTxO -> Tx -> property) -> Property

forAllCommit :: Testable property => (UTxO -> Tx -> property) -> Property

forAllCommit' :: Testable property => (ChainContext -> InitialState -> UTxO -> Tx -> property) -> Property

forAllAbort :: Testable property => (UTxO -> Tx -> property) -> Property

forAllCollectCom :: Testable property => (UTxO -> Tx -> property) -> Property

forAllDeposit :: Testable property => (UTxO -> Tx -> property) -> Property

forAllRecover :: Testable property => (UTxO -> Tx -> property) -> Property

forAllIncrement :: Testable property => (UTxO -> Tx -> property) -> Property

forAllIncrement' :: Testable property => (UTxO -> Tx -> property) -> Property

forAllDecrement :: Testable property => (UTxO -> Tx -> property) -> Property

forAllDecrement' :: Testable property => (UTxO -> UTxO -> Tx -> property) -> Property

forAllClose :: Testable property => (UTxO -> Tx -> property) -> Property

forAllContest :: Testable property => (UTxO -> Tx -> property) -> Property

forAllFanout :: Testable property => (UTxO -> Tx -> property) -> Property

genInitTxWithSeed :: HydraContext -> Gen (Tx, TxIn)

Generate an init tx with the used seed TxIn.

Helpers

mfail :: MonadFail m => Maybe a -> m a