Hydra.Chain.Direct.TxSpec

Description

Unit tests for our "hand-rolled" transactions as they are used in the "direct" chain component.

Documentation

spec :: Spec

ledgerPParams :: BabbagePParams LedgerEra

withinTxExecutionBudget :: EvaluationReport -> Property

generateCommitUTxOs :: [Party] -> Gen (Map TxIn (TxOut CtxUTxO, HashableScriptData, UTxO))

Generate a UTXO representing commit outputs for a given list of Party. NOTE: Uses testPolicyId for the datum. NOTE: We don't generate empty commits and it is used only at one place so perhaps move it? FIXME: This function is very complicated and it's hard to understand it after a while

prettyEvaluationReport :: EvaluationReport -> String

genAbortableOutputs :: [Party] -> Gen ([UTxOWithScript], [(TxIn, TxOut CtxUTxO, HashableScriptData, UTxO)])

fst3 :: (a, b, c) -> a

third :: (a, b, c) -> c

drop2nd :: (a, b, c) -> (a, c)

drop3rd :: (a, b, c) -> (a, b)

tripleToPair :: (a, b, c) -> (a, (b, c))