Hydra.Chain.Direct.TxTraceSpec

Documentation

spec :: Spec

prop_traces :: Property

prop_runActions :: Actions Model -> Property

Model

data Model

Constructors

Model

Fields

snapshots :: [SnapshotNumber]

headState :: State

utxoV :: Var UTxO

Last known, spendable UTxO.

alreadyContested :: [Actor]

data State

Constructors

Open

Closed

Final

data Actor

Constructors

Alice

Bob

Carol

Fixtures and glue code

allActors :: [Actor]

List of all model actors corresponding to the fixtures used.

snapshotUTxO :: SnapshotNumber -> UTxO

A "random" UTxO distribution for a given snapshot number. This always contains one UTxO for alice, bob, and carol.

correctlySignedSnapshot :: SnapshotNumber -> ConfirmedSnapshot Tx

A model of a correctly signed snapshot. Given a snapshot number a snapshot signed by all participants (alice, bob and carol) with some UTxO contained is produced.

openHeadUTxO :: UTxO

UTxO of the open head on-chain. NOTE: This uses fixtures for headId, parties, and cperiod.

newCloseTx :: HasCallStack => Actor -> ConfirmedSnapshot Tx -> IO Tx

Creates a transaction that closes openHeadUTxO with given the snapshot. NOTE: This uses fixtures for headId, parties (alice, bob, carol), contestation period and also claims to close at time 0 resulting in a contestation deadline of 0 + cperiod.

newContestTx :: HasCallStack => UTxO -> Actor -> ConfirmedSnapshot Tx -> IO Tx

Creates a contest transaction using given utxo and contesting with given snapshot. NOTE: This uses fixtures for headId, contestation period and also claims to contest at time 0.

actorChainContext :: Actor -> ChainContext

Fixture for the chain context of a model Actor on testNetworkId. Uses a generated ScriptRegistry.

testScriptRegistry :: ScriptRegistry

Helpers

validateTx :: (HasCallStack, MonadThrow m) => UTxO -> Tx -> m ()

Thin wrapper around evaluateTx that fails with failure if any of the scripts/redeemers fail to evaluate.

observeTxMatching :: (HasCallStack, MonadThrow m) => UTxO -> Tx -> (HeadObservation -> Maybe a) -> m a

Expect to observe a transaction matching given predicate. This fails with failure if the predicate yields Nothing.