Hydra.Chain.Direct.Wallet

Description

Companion tiny-wallet for the direct chain component. This module provide some useful utilities to tracking the wallet's UTXO, and accessing it

Documentation

type Address = Addr

type TxIn = TxIn

type TxOut = TxOut LedgerEra

data TinyWallet m

A TinyWallet is a small abstraction of a wallet with basic UTXO management. The wallet is assumed to have only one address, and only one UTXO at that address. It can sign transactions and keeps track of its UTXO behind the scene.

The wallet is connecting to the node initially and when asked to $sel:reset:TinyWallet. Otherwise it can be fed blocks via $sel:update:TinyWallet as the chain rolls forward.

Constructors

TinyWallet

Fields

getUTxO :: STM m (Map TxIn TxOut)

Return all known UTxO addressed to this wallet.

getSeedInput :: STM m (Maybe TxIn)

Returns the seed input This is the special input needed by Direct chain component to initialise a head

sign :: Tx -> Tx

coverFee :: UTxO -> Tx -> m (Either ErrCoverFee Tx)

reset :: m ()

Re-initializ wallet against the latest tip of the node and start to ignore $sel:update:TinyWallet calls until reaching that tip.

update :: BlockHeader -> [Tx] -> m ()

Update the wallet state given a block and list of txs. May be ignored if wallet is still initializing.

data WalletInfoOnChain

Constructors

WalletInfoOnChain

Fields

walletUTxO :: Map TxIn TxOut

systemStart :: SystemStart

tip :: ChainPoint

Latest point on chain the wallet knows of.

type ChainQuery m = QueryPoint -> Address ShelleyAddr -> m WalletInfoOnChain

newTinyWallet

Arguments

:: Tracer IO TinyWalletLog
A tracer for logging

-> NetworkId
Network identifier to generate our address.

-> (VerificationKey PaymentKey, SigningKey PaymentKey)
Credentials of the wallet.

-> ChainQuery IO
A function to query UTxO, pparams, system start and epoch info from the node. Initially and on demand later.

-> IO (EpochInfo (Either Text))

-> IO (PParams ConwayEra)
A means to query some pparams.

-> IO (TinyWallet IO)

Create a new tiny wallet handle.

applyTxs :: [Tx] -> (Address -> Bool) -> Map TxIn TxOut -> Map TxIn TxOut

Apply a block to our wallet. Does nothing if the transaction does not modify the UTXO set, or else, remove consumed utxos and add produced ones.

To determine whether a produced output is ours, we apply the given function checking the output's address.

getTxId :: HashAnnotated (TxBody era) EraIndependentTxBody => AlonzoTx era -> TxId

data ErrCoverFee

This are all the error that can happen during coverFee.

Constructors

ErrNotEnoughFunds ChangeError

ErrNoFuelUTxOFound

ErrUnknownInput

Fields

input :: TxIn

ErrScriptExecutionFailed

Fields

redeemerPointer :: Text

scriptFailure :: Text

ErrTranslationError (ContextError LedgerEra)

ErrConwayUpgradeError (TxUpgradeError ConwayEra)

data ChangeError

Constructors

ChangeError

Fields

inputBalance :: Coin

outputBalance :: Coin

coverFee_ :: forall era. (EraPlutusContext era, EraCertState era, AlonzoEraTx era, ScriptsNeeded era ~ AlonzoScriptsNeeded era, EraUTxO era, BabbageEraTxBody era) => PParams era -> SystemStart -> EpochInfo (Either Text) -> Map TxIn (TxOut era) -> Map TxIn (TxOut era) -> Tx era -> Either ErrCoverFee (Tx era)

Cover fee for a transaction body using the given UTXO set. This calculate necessary fees and augments inputs outputs collateral accordingly to cover for the transaction cost and get the change back.

XXX: All call sites of this function use cardano-api types

findLargestUTxO :: EraTxOut era => Map TxIn (TxOut era) -> Maybe (TxIn, TxOut era)

estimateScriptsCost

Arguments

:: forall era. (AlonzoEraTx era, EraPlutusContext era, ScriptsNeeded era ~ AlonzoScriptsNeeded era, EraUTxO era)

=> PParams era
Protocol parameters

-> SystemStart
Start of the blockchain, for converting slots to UTC times

-> EpochInfo (Either Text)
Information about epoch sizes, for converting slots to UTC times

-> Map TxIn (TxOut era)
A UTXO needed to resolve inputs

-> Tx era
The pre-constructed transaction

-> Either ErrCoverFee (Map (PlutusPurpose AsIx era) ExUnits)

Estimate cost of script executions on the transaction. This is only an estimates because the transaction isn't sealed at this point and adding new elements to it like change outputs or script integrity hash may increase that cost a little.

data TinyWalletLog

Constructors

BeginInitialize

EndInitialize

Fields

initialUTxO :: UTxO

tip :: ChainPoint

BeginUpdate

Fields

point :: ChainPoint

EndUpdate

Fields

newUTxO :: UTxO

SkipUpdate

Fields

point :: ChainPoint