Hydra.Client

Documentation

data HydraEvent tx

Constructors

ClientConnected

ClientDisconnected

Update (AllPossibleAPIMessages tx)

Tick UTCTime

data AllPossibleAPIMessages tx

All possible messages that expect to receive from the hydra-node.

Constructors

ApiTimedServerOutput (TimedServerOutput tx)

ApiClientMessage (ClientMessage tx)

ApiGreetings (Greetings tx)

ApiInvalidInput InvalidInput

data Client tx m

Handle to interact with Hydra node

Constructors

Client

Fields

sendInput :: ClientInput tx -> m ()

Send some input to the server.

sk :: SigningKey PaymentKey

externalCommit :: UTxO -> m ()

recoverCommit :: TxId -> m ()

type ClientCallback tx m = HydraEvent tx -> m ()

Callback for receiving server outputs.

type ClientComponent tx m a = ClientCallback tx m -> (Client tx m -> m a) -> m a

A type tying both receiving output and sending input into a Component.

withClient :: forall tx a. IsChainState tx => Options -> ClientComponent tx IO a

Provide a component to interact with Hydra node.

data ClientError

Constructors

ClientJSONDecodeError String ByteString