Hydra.Cluster.Faucet

Documentation

data FaucetException

Constructors

FaucetHasNotEnoughFunds

Fields

faucetUTxO :: UTxO

FaucetFailedToBuildTx

Fields

reason :: TxBodyErrorAutoBalance Era

FaucetBlockfrostError

Fields

blockFrostError :: Text

data FaucetLog

Constructors

TraceResourceExhaustedHandled Text

ReturnedFunds

Fields

returnAmount :: Coin

delayBF :: (MonadDelay m, ChainBackend backend) => backend -> m ()

seedFromFaucet

Arguments

:: ChainBackend backend

=> backend

-> VerificationKey PaymentKey
Recipient of the funds

-> Value
Value to get from faucet

-> Tracer IO FaucetLog

-> IO UTxO

seedFromFaucetWithMinting

Arguments

:: ChainBackend backend

=> backend

-> VerificationKey PaymentKey
Recipient of the funds

-> Value
Value to get from faucet

-> Tracer IO FaucetLog

-> Maybe PlutusScript

-> IO UTxO

Create a specially marked "seed" UTXO containing requested Value by redeeming funds available to the well-known faucet.

findFaucetUTxO :: ChainBackend backend => NetworkId -> backend -> Coin -> IO UTxO

seedFromFaucetBlockfrost

Arguments

:: BlockfrostOptions

-> VerificationKey PaymentKey
Recipient of the funds

-> Coin
Amount to get from faucet

-> BlockfrostClientT IO UTxO

seedFromFaucet_

Arguments

:: ChainBackend backend

=> backend

-> VerificationKey PaymentKey
Recipient of the funds

-> Coin
Amount to get from faucet

-> Tracer IO FaucetLog

-> IO ()

Like seedFromFaucet, but without returning the seeded UTxO.

returnFundsToFaucet :: ChainBackend backend => Tracer IO FaucetLog -> backend -> Actor -> IO ()

Return the remaining funds to the faucet

returnFundsToFaucet' :: ChainBackend backend => Tracer IO FaucetLog -> backend -> SigningKey PaymentKey -> IO Coin

createOutputAtAddress :: ChainBackend backend => NetworkId -> backend -> AddressInEra -> TxOutDatum CtxTx -> Value -> IO (TxIn, TxOut CtxUTxO)

retryOnExceptions :: (MonadCatch m, MonadDelay m, ChainBackend backend) => Tracer m FaucetLog -> backend -> m a -> m a

Try to submit tx and retry when some caught exception/s take place.

publishHydraScriptsAs :: ChainBackend backend => backend -> Actor -> IO [TxId]

Publish current Hydra scripts as scripts outputs for later referencing them.

The key of the given Actor is used to pay for fees in required transactions, it is expected to have sufficient funds.