Hydra.Cluster.Fixture

Description

Example values and utilities used across hydra-cluster tests and benchmarks.

Documentation

alice :: Party

bob :: Party

carol :: Party

aliceSk :: SigningKey HydraKey

bobSk :: SigningKey HydraKey

carolSk :: SigningKey HydraKey

aliceVk :: VerificationKey HydraKey

bobVk :: VerificationKey HydraKey

carolVk :: VerificationKey HydraKey

cperiod :: ContestationPeriod

blockfrostcperiod :: ContestationPeriod

availableInitialFunds :: Num a => a

data Actor

Enumeration of known actors for which we can get the keysFor and writeKeysFor.

Constructors

Alice

AliceFunds

Bob

BobFunds

Carol

CarolFunds

Faucet

actorName :: Actor -> String

fundsOf :: Actor -> Actor

data KnownNetwork

A network known to the hydra-cluster. That means we have configuration files to connect to at least these networks.

Constructors

Preview

Preproduction

Mainnet

Sanchonet

BlockfrostPreview

BlockfrostPreprod

BlockfrostMainnet

toNetworkId :: KnownNetwork -> NetworkId