Hydra.Cluster.Util

Description

Utilities used across hydra-cluster

Documentation

readConfigFile :: FilePath -> IO ByteString

Lookup a config file similar reading a file from disk. If the env variable HYDRA_CONFIG_DIR is set, filenames will be resolved relative to its value otherwise they will be looked up in the package's data path.

keysFor :: Actor -> IO (VerificationKey PaymentKey, SigningKey PaymentKey)

Get the "well-known" keys for given actor.

createAndSaveSigningKey :: FilePath -> IO (SigningKey PaymentKey)

Create and save new signing key at the provided path. NOTE: Uses TextEnvelope format.

type BlockTime = NominalDiffTime

Expected time between blocks (on average)

data Timing

Timing parameters that determine the behavior of a (cluster of) hydra-node.

Constructors

Timing

Fields

blockTime :: BlockTime

contestationPeriod :: ContestationPeriod

depositPeriod :: DepositPeriod

mkTestTiming :: BlockTime -> Timing

Set up reasonable timing parameters for testing given a BlockTime.

mkTestTiming' :: Int -> BlockTime -> Timing

Like mkTestTiming but scales Timing by the number of concurrent deposits expected. Each increment tx must be processed sequentially on-chain, so N concurrent deposits require N times the base deposit period.

depositTimeout :: Timing -> NominalDiffTime

Get a timeout until a deposit should have happened given a Timing.

chainConfigFor

Arguments

:: ChainBackend backend

=> HasCallStack

=> Actor

-> FilePath

-> backend

-> [TxId]
Transaction ids at which Hydra scripts should have been published.

-> [Actor]

-> Timing

-> IO ChainConfig

Create a (test) chain config for a given actor.

chainConfigFor'

Arguments

:: ChainBackend backend

=> HasCallStack

=> Actor

-> FilePath

-> backend

-> [TxId]
Transaction ids at which Hydra scripts should have been published.

-> [Actor]

-> ContestationPeriod

-> DepositPeriod

-> IO ChainConfig

modifyConfig :: (CardanoChainConfig -> CardanoChainConfig) -> ChainConfig -> ChainConfig

setNetworkId :: NetworkId -> ChainConfig -> ChainConfig