Hydra.Contract.HeadState

Documentation

type SnapshotNumber = Integer

type SnapshotVersion = Integer

type Hash = BuiltinByteString

type Signature = BuiltinByteString

data OpenDatum

Sub-type for the open state-machine state.

Constructors

OpenDatum

Fields

headSeed :: TxOutRef

TODO: Spec?

headId :: CurrencySymbol

Spec: cid

parties :: [Party]

Spec: kH

contestationPeriod :: ContestationPeriod

Spec: T

version :: SnapshotVersion

Spec: v

accumulatorHash :: Hash

Digest of the accumulator hash for the last confirmed snapshot

headAdaOverhead :: Integer

Lovelace in the head UTxO not belonging to any L2 UTxO (min-UTxO overhead). Set once at init time and invariant for the head's lifetime.

data ClosedDatum

Sub-type for the closed state-machine state.

Constructors

ClosedDatum

Fields

headId :: CurrencySymbol

Spec: cid

parties :: [Party]

Spec: kH

contestationPeriod :: ContestationPeriod

Spec: T

version :: SnapshotVersion

Spec: v

snapshotNumber :: SnapshotNumber

Spec: s

contesters :: [PubKeyHash]

Spec: C

contestationDeadline :: POSIXTime

Spec: tfinal

accumulatorCommitment :: BuiltinBLS12_381_G1_Element

KZG commitment to the full UTxO set.

headAdaOverhead :: Integer

Lovelace in the head UTxO not belonging to any L2 UTxO (min-UTxO overhead). Propagated unchanged from OpenDatum via Close.

data FanoutProgressDatum

Sub-type for intermediate partial fanout state. Carries only the fields needed for subsequent partial fanout steps.

Constructors

FanoutProgressDatum

Fields

headId :: CurrencySymbol

parties :: [Party]

contestationDeadline :: POSIXTime

accumulatorCommitment :: BuiltinBLS12_381_G1_Element

headAdaOverhead :: Integer

Lovelace in the head UTxO not belonging to any L2 UTxO. Propagated from ClosedDatum.

progressFromClosed :: ClosedDatum -> FanoutProgressDatum

Extract the fields needed for partial fanout steps from a ClosedDatum. Called both on-chain (in the validator dispatch) and off-chain (in tx building).

data State

Constructors

Open OpenDatum

Closed ClosedDatum

Final

FanoutProgress FanoutProgressDatum

data CloseRedeemer

Sub-type for close transition with auxiliary data as needed.

Constructors

CloseInitial
Initial snapshot is used to close.

CloseAny
Any snapshot which doesn't contain anything to inc/decrement but snapshot number is higher than zero.

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

accumulatorHash :: Hash

Digest of the accumulator hash

CloseUnused
Closing snapshot refers to the current state version (pending inc/dec not yet applied)

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

accumulatorHash :: Hash

Digest of the accumulator hash

CloseUsed
Closing snapshot refers to the previous state version (pending inc/dec already applied)

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

accumulatorHash :: Hash

Digest of the accumulator hash

data ContestRedeemer

Sub-type for contest transition with auxiliary data as needed.

Constructors

ContestUnused
Contesting snapshot refers to the current state version (inc/dec not yet applied, or no pending action)

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

accumulatorHash :: Hash

Digest of the accumulator hash

ContestUsed
Contesting snapshot refers to the previous state version (pending inc/dec already applied)

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

accumulatorHash :: Hash

Digest of the accumulator hash

data IncrementRedeemer

Sub-type for increment transition

Constructors

IncrementRedeemer

Fields

signature :: [Signature]

snapshotNumber :: SnapshotNumber

increment :: TxOutRef

data DecrementRedeemer

Sub-type for decrement transition with auxiliary data as needed.

Constructors

DecrementRedeemer

Fields

signature :: [Signature]

Spec: ξ

snapshotNumber :: SnapshotNumber

Spec: s

numberOfDecommitOutputs :: Integer

Spec: m

data Input

Constructors

Increment IncrementRedeemer

Decrement DecrementRedeemer

Close CloseRedeemer

Contest ContestRedeemer

Fanout

Fields

numberOfFanoutOutputs :: Integer

proof :: BuiltinBLS12_381_G1_Element

crsRef :: TxOutRef

PartialFanout

Fields

numberOfPartialOutputs :: Integer

crsRef :: TxOutRef

FinalPartialFanout

Fields

numberOfPartialOutputs :: Integer

proof :: BuiltinBLS12_381_G1_Element

crsRef :: TxOutRef