Hydra.Data.Party

Documentation

newtype Party

On-chain representation of a Hydra party.

NOTE: This roughly corresponds to the Party in 'hydra-node', but is simplified to allow usage of this type in plutus-tx. If we would use the complex type directly, which is based on 'cardano-crypto-class', we would get errors like "Error: Unsupported feature: Kind: GHC.Types.Nat".

The data constructor should not be used to construct this value as it would always come from off-chain code via partyFromVerificationKeyBytes.

Constructors

UnsafeParty

Fields

vkey :: BuiltinByteString

partyFromVerificationKeyBytes :: ByteString -> Party

Create an on-chain Party from some verification key bytes.

partyToVerificationKeyBytes :: Party -> ByteString

Get the verification key bytes contained from an on-chain Party.