Hydra.Events

Description

This module defines the types and functions for creating EventSource and EventSink instances and is intended to be used as an extension point.

A single EventSource and zero or more EventSink handles are used by the main HydraNode handle to load and send out events.

See FileBased for an example implementation and FileBasedSpec for the corresponding test suite.

Custom implementations should be located under Hydra.Events to avoid conflicts.

Documentation

type EventId = Word64

class HasEventId a where

Methods

getEventId :: a -> EventId

newtype EventSource e m

Constructors

EventSource

Fields

sourceEvents :: HasEventId e => ConduitT () e (ResourceT m) ()

Stream all events from the event source.

getEvents :: (HasEventId e, MonadUnliftIO m) => EventSource e m -> m [e]

Retrieve all events from the event source as a list.

newtype EventSink e m

Constructors

EventSink

Fields

putEvent :: HasEventId e => e -> m ()

Send a single event to the event sink.

putEventsToSinks :: (Monad m, HasEventId e) => [EventSink e m] -> [e] -> m ()

Put a list of events to a list of event sinks in a round-robin fashion.