Hydra.Generator

Documentation

networkId :: NetworkId

data Dataset

A Dataset that can be run for testing purpose. Each Dataset represents a complete scenario where several ClientDataset are run concurrently against one or more HydraNodes. A dataset can optionally have a title and description which will be used to report results.

Constructors

Dataset

Fields

fundingTransaction :: Tx

hydraNodeKeys :: [SigningKey PaymentKey]

Cardano signing keys that will hold fuel.

clientDatasets :: [ClientDataset]

title :: Maybe Text

description :: Maybe Text

data ClientDataset

Constructors

ClientDataset

Fields

paymentKey :: SigningKey PaymentKey

initialUTxO :: UTxO

txSequence :: [Tx]

generateConstantUTxODataset

Arguments

:: SigningKey PaymentKey
Faucet signing key

-> Int
Number of clients

-> Int
Number of transactions

-> Gen Dataset

Generate a Dataset which does not grow the per-client UTXO set over time. This version provided faucet key owns funds on the initial funds of the devnet (See availableInitialFunds and 'genesis-shelley.json'). Then for a given number of clients a number of transactions are generated.

generateDemoUTxODataset

Arguments

:: NetworkId

-> SocketPath

-> SigningKey PaymentKey
Faucet signing key

-> Int
Number of clients.

-> Int
Number of transactions

-> IO Dataset

Generate a Dataset from an already running network by querying available funds of the well-known 'faucet.sk' and assuming the hydra-nodes we connect to have fuel available. Then for a given number of clients a number of transactions are generated.

Helpers

withInitialUTxO :: SigningKey PaymentKey -> Tx -> UTxO

genClientFunds :: [SigningKey PaymentKey] -> Integer -> Gen [(VerificationKey PaymentKey, Coin)]

generateClientDataset :: NetworkId -> Tx -> Int -> SigningKey PaymentKey -> Gen ClientDataset