Hydra.HeadLogic.Input

Documentation

type TTL = Natural

data Input tx

Inputs that are processed by the head logic (the "core"). Corresponding to each of the "shell" layers, we distinguish between inputs from the client, the network and the chain.

Constructors

ClientInput
Input received from clients via the Hydra.API.

Fields

clientInput :: ClientInput tx

NetworkInput
Input received from peers via a Hydra.Network.

$sel:ttl:ClientInput is a simple counter that's decreased every time the event is reenqueued due to a wait. It's default value is defaultTTL

Fields

ttl :: TTL

networkEvent :: NetworkEvent (Message tx)

ChainInput
Input received from the chain via a Hydra.Chain.

Fields

chainEvent :: ChainEvent tx