Hydra.HeadLogicSpec

Description

Unit tests of the the protocol logic in HeadLogic. These are very fine grained and specific to individual steps in the protocol. More high-level of the protocol logic, especially between multiple parties can be found in BehaviorSpec.

Documentation

spec :: Spec

runEvents :: Monad m => Environment -> Ledger tx -> HeadState tx -> StateT (StepState tx) m a -> m a

chainEffect :: PostChainTx SimpleTx -> Effect SimpleTx

Create a chain effect with fixed chain state and slot.

observeEventAtSlot :: Natural -> OnChainTx SimpleTx -> Event SimpleTx

observationEvent :: OnChainTx SimpleTx -> Event SimpleTx

Create an observation event with fixed chain state and slot.

inInitialState :: [Party] -> HeadState SimpleTx

inOpenState :: [Party] -> Ledger SimpleTx -> HeadState SimpleTx

inOpenState' :: [Party] -> CoordinatedHeadState SimpleTx -> HeadState SimpleTx

inClosedState :: [Party] -> HeadState SimpleTx

inClosedState' :: [Party] -> ConfirmedSnapshot SimpleTx -> HeadState SimpleTx

getConfirmedSnapshot :: HeadState tx -> Maybe (Snapshot tx)

data StepState tx

Constructors

StepState

Fields

headState :: HeadState tx

env :: Environment

ledger :: Ledger tx

step :: (MonadState (StepState tx) m, HasCallStack, IsChainState tx) => Event tx -> m (HeadState tx)

Asserts that the update function will update the state (return a NewState) for this Event

assertUpdateState :: (MonadState (HeadState tx) m, HasCallStack, IsChainState tx) => Environment -> Ledger tx -> Event tx -> m (HeadState tx)

Asserts that the update function will update the state (return a NewState) for this Event

assertNewState :: (HasCallStack, IsChainState tx, Monad m) => Outcome tx -> m (HeadState tx)

assertEffects :: (HasCallStack, IsChainState tx) => Outcome tx -> IO ()

hasEffect :: (HasCallStack, IsChainState tx) => Outcome tx -> Effect tx -> IO ()

hasWait :: (HasCallStack, IsChainState tx) => Outcome tx -> WaitReason tx -> IO ()

hasEffectSatisfying :: (HasCallStack, IsChainState tx) => Outcome tx -> (Effect tx -> Bool) -> IO ()

hasNoEffectSatisfying :: (HasCallStack, IsChainState tx) => Outcome tx -> (Effect tx -> Bool) -> IO ()

testHeadId :: HeadId