Hydra.Ledger

Documentation

nextChainSlot :: ChainSlot -> ChainSlot

Get the next chain slot. Use this instead of giving Enum or Num instances to ChainSlot.

newtype Ledger tx

An abstract interface for a Ledger. Allows to define mock / simpler implementation for testing as well as limiting feature-envy from the business logic by forcing a closed interface.

Constructors

Ledger

Fields

applyTransactions :: ChainSlot -> UTxOType tx -> [tx] -> Either (tx, ValidationError) (UTxOType tx)

Apply a set of transaction to a given UTxO set. Returns the new UTxO or validation failures returned from the ledger. TODO: ValidationError should also include the UTxO, which is not necessarily the same as the given UTxO after some transactions

collectTransactions :: Ledger tx -> ChainSlot -> UTxOType tx -> [tx] -> ([tx], UTxOType tx)

Collect applicable transactions and resulting UTxO. In contrast to $sel:applyTransactions:Ledger, this functions continues on validation errors.

data ValidationResult

Either valid or an error which we get from the ledger-specs tx validation.

Constructors

Valid

Invalid ValidationError

newtype ValidationError

Constructors

ValidationError

Fields

reason :: Text