Hydra.Ledger.Cardano.Builder

Description

Utilities to building transactions on top of the cardano-api.

Executing

unsafeBuildTransaction :: HasCallStack => TxBodyContent BuildTx -> Tx

Construct a transaction from a builder. It is said unsafe because the underlying implementation will perform some sanity check on a transaction; for example, check that it has at least one input, that no outputs are negatives and whatnot.

We use the builder only internally for on-chain transaction crafted in the context of Hydra.

data InvalidTransactionException

A runtime exception to capture (programmer) failures when building transactions. This should never happened in practice (famous last words...)!

Constructors

InvalidTransactionException

Fields

txBodyError :: TxBodyError

builder :: TxBodyContent BuildTx

addTxInsSpending :: [TxIn] -> TxBodyContent BuildTx -> TxBodyContent BuildTx

Like addInputs but only for vk inputs which requires no additional data.

mintTokens :: ToScriptData redeemer => PlutusScript -> redeemer -> PolicyAssets -> TxBodyContent BuildTx -> TxBodyContent BuildTx

Mint tokens with given plutus minting script and redeemer.

burnTokens :: ToScriptData redeemer => PlutusScript -> redeemer -> PolicyAssets -> TxBodyContent BuildTx -> TxBodyContent BuildTx

Burn tokens with given plutus minting script and redeemer. This is really just mintTokens with negated Quantity.