Hydra.Ledger.CardanoSpec

Documentation

spec :: Spec

shouldParseJSONAs :: forall a. (HasCallStack, FromJSON a) => LByteString -> Expectation

roundtripPParams :: PParams LedgerEra -> Property

Test that the PParams To/FromJSON instances to roundtrip.

appliesValidTransaction :: Property

appliesValidTransactionFromJSON :: Property

propRealisticValue :: Value -> Property

A transaction or transaction output can usually only contain a realistic number of native asset entries. This property checks a realistic order of magnitude (100).

propHasEnoughLovelace :: TxOut CtxUTxO -> Property

Check that an output has enough lovelace to cover asset deposits.

propGeneratesGoodTxOut :: TxOut CtxUTxO -> Property

Check that the given TxOut fulfills several requirements and does not use unsupported features. See genTxOut for rationale.