Hydra.Logging

Description

Adapter module to the actual logging framework. All Hydra node components implements Structured logging via contra-tracer generic logging framework. All logs are output in JSON.

Tracer

newtype Tracer (m :: Type -> Type) a

Constructors

Tracer

Fields

runTracer :: a -> m ()

natTracer :: (forall x. m x -> n x) -> Tracer m s -> Tracer n s

nullTracer :: forall (m :: Type -> Type) a. Applicative m => Tracer m a

traceWith :: Tracer m a -> a -> m ()

class ToObject a where

Minimal complete definition

Nothing

Methods

toObject :: TracingVerbosity -> a -> Object

textTransformer :: a -> Object -> Text

data TracingVerbosity

Constructors

MinimalVerbosity

NormalVerbosity

MaximalVerbosity

Using it

data Verbosity

Constructors

Quiet

Verbose Text

data Envelope a

Provides logging metadata for entries.

Constructors

Envelope

Fields

timestamp :: UTCTime

threadId :: Int

namespace :: Text

message :: a

withTracer :: forall m msg a. (MonadIO m, MonadFork m, MonadTime m, ToJSON msg) => Verbosity -> (Tracer m msg -> IO a) -> IO a

Start logging thread and acquire a Tracer. This tracer will dump all messages on stdout, one message per line, formatted as JSON. This tracer is wrapping msg into an Envelope with metadata.

withTracerOutputTo :: forall m msg a. (MonadIO m, MonadFork m, MonadTime m, ToJSON msg) => Handle -> Text -> (Tracer m msg -> IO a) -> IO a

Start logging thread acquiring a Tracer, outputting JSON formatted messages to some Handle. This tracer is wrapping msg into an Envelope with metadata.

showLogsOnFailure :: (MonadLabelledSTM m, MonadCatch m, MonadFork m, MonadTime m, MonadSay m, ToJSON msg) => Text -> (Tracer m msg -> m a) -> m a

Capture logs and output them to stdout when an exception was raised by the given action. This tracer is wrapping msg into an Envelope with metadata.

traceInTVar :: (MonadFork m, MonadTime m, MonadSTM m) => TVar m [Envelope msg] -> Text -> Tracer m msg

contramap :: Contravariant f => (a' -> a) -> f a -> f a'