Hydra.Model.Payment

Description

A simplistic type of transactions useful for modelling purpose. a Payment is a simple transaction type that moves some amount of ADAs between to CardanoSigningKey.

Documentation

newtype CardanoSigningKey

Constructors

CardanoSigningKey

Fields

signingKey :: SigningKey PaymentKey

data Payment

A single Ada-payment only transaction in our model.

Constructors

Payment

Fields

from :: CardanoSigningKey

to :: CardanoSigningKey

value :: Value

applyTx :: UTxOType Payment -> Payment -> UTxOType Payment

genAdaValue :: Gen Value

Orphans

Orphan instances

Arbitrary Value

Instance details

Methods

arbitrary :: Gen Value

shrink :: Value -> [Value]

FromCBOR Value

Instance details

Methods

fromCBOR :: Decoder s Value

label :: Proxy Value -> Text

ToCBOR Value

Instance details

Methods

toCBOR :: Value -> Encoding

encodedSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy Value -> Size

encodedListSizeExpr :: (forall t. ToCBOR t => Proxy t -> Size) -> Proxy [Value] -> Size