Hydra.Network.Authenticate

Description

A NetworkComponent that handles authentication of sent and received messages.

This "middleware" uses HydraKey keys for signing own messages and verifying others', providing Authenticated messages to consumers.

Documentation

data Signed msg

Represents a signed message over the network. Becomes valid once its receivers verify it against its other peers verification keys. Messages are signed and turned into authenticated messages before broadcasting them to other peers.

Constructors

Signed

Fields

payload :: msg

signature :: Signature msg

party :: Party

data Authenticated msg

Constructors

Authenticated

Fields

payload :: msg

party :: Party

withAuthentication :: (SignableRepresentation inbound, ToJSON inbound, SignableRepresentation outbound) => Tracer m AuthLog -> SigningKey HydraKey -> [Party] -> NetworkComponent m (Signed inbound) (Signed outbound) a -> NetworkComponent m (Authenticated inbound) outbound a

Middleware used to sign messages before broadcasting them to other peers and verify signed messages upon receiving. Only verified messages are pushed downstream to the internal network for the node to consume and process. Non-verified messages get discarded.

mkAuthLog :: (ToJSON msg, Show signature) => msg -> signature -> Party -> AuthLog

Smart constructor for MessageDropped

data AuthLog

Constructors

MessageDropped

Fields

message :: Text

signature :: Text

party :: Party