Hydra.Node.ParameterMismatch

Description

Structured errors related to configuration mismatch.

When we start a Node we need to do sanity checks between what's provided as parameters to the node and what's persisted.

Documentation

newtype ParameterMismatch

Exception used to indicate command line options not matching the persisted state.

Constructors

ParameterMismatch [ParamMismatch]

data ParamMismatch

Constructors

ContestationPeriodMismatch

Fields

loadedCp :: ContestationPeriod

configuredCp :: ContestationPeriod

PartiesMismatch

Fields

loadedParties :: [Party]

configuredParties :: [Party]

SavedNetworkPartiesInconsistent

Fields

numberOfParties :: Int