Hydra.Node.UnsyncedPeriod

Documentation

newtype UnsyncedPeriod

Period of time after which we consider the node becoming unsynced with the chain. Beyond this period the node will refuse to process new transactions and signing snapshots.

Constructors

UnsyncedPeriod

Fields

unsyncedPeriodToNominalDiffTime :: NominalDiffTime

defaultUnsyncedPeriodFor :: ContestationPeriod -> UnsyncedPeriod

Compute a default UnsyncedPeriod based on the ContestationPeriod. This is the legacy behavior: half of the contestation period.