Hydra.NodeSpec

Documentation

spec :: Spec

primeWith :: (MonadSTM m, MonadTime m) => [Input tx] -> HydraNode tx m -> m (HydraNode tx m)

Add given list of inputs to the InputQueue. A preceding Tick is enqueued to advance the chain slot and ensure the NodeState is in sync. This is returning the node to allow for chaining with runToCompletion.

notConnect :: MonadThrow m => DraftHydraNode tx m -> m (HydraNode tx m)

Convert a DraftHydraNode to a HydraNode by providing mock implementations.

mockServer :: Monad m => Server tx m

mockNetwork :: Monad m => Network m (Message tx)

mockChain :: MonadThrow m => Chain tx m

mockSink :: Monad m => EventSink a m

mockEventStore :: forall a m. Monad m => [a] -> EventStore a m

mockSource :: Monad m => [a] -> EventSource a m

createRecordingSink :: IO (EventSink a IO, IO [a])

createMockEventStore :: MonadLabelledSTM m => m (EventStore a m)

inputsToOpenHead :: [Input SimpleTx]

observationInput :: OnChainTx SimpleTx -> Input SimpleTx

runToCompletion :: IsChainState tx => HydraNode tx IO -> IO ()

testHydraNode :: (MonadTime m, MonadDelay m, MonadAsync m, MonadLabelledSTM m, MonadThrow m, MonadUnliftIO m) => Tracer m (HydraNodeLog SimpleTx) -> SigningKey HydraKey -> [Party] -> ContestationPeriod -> [Input SimpleTx] -> m (HydraNode SimpleTx m)

Creates a full HydraNode with given parameters and primed Inputs. Note that this node is notConnected to any components.

testTTL :: TTL

A lot shorter than the real world ttl which requires to be long to overcome some longer chain observation delays, but would just make tests longer.

recordNetwork :: HydraNode tx IO -> IO (HydraNode tx IO, IO [Message tx])

recordServerOutputs :: IsChainState tx => HydraNode tx IO -> IO (HydraNode tx IO, IO [Either (ServerOutput tx) (ClientMessage tx)])

messageRecorder :: IO (msg -> IO (), IO [msg])

throwExceptionOnPostTx :: IsChainState tx => PostTxError tx -> HydraNode tx IO -> IO (HydraNode tx IO)