Hydra.Options

Documentation

data Command

Constructors

Run RunOptions

Publish PublishOptions

GenHydraKey GenerateKeyPair

data ChainBackendOptions

Constructors

Direct DirectOptions

Blockfrost BlockfrostOptions

data BlockfrostOptions

Constructors

BlockfrostOptions

Fields

projectPath :: FilePath

Path to the blockfrost project file

queryTimeout :: Int

retryTimeout :: Int

data CardanoChainConfig

Constructors

CardanoChainConfig

Fields

hydraScriptsTxId :: [TxId]

Identifier of transaction holding the hydra scripts to use.

cardanoSigningKey :: FilePath

Path to the cardano signing key of the internal wallet.

cardanoVerificationKeys :: [FilePath]

Paths to other node's verification keys.

startChainFrom :: Maybe ChainPoint

Point at which to start following the chain.

contestationPeriod :: ContestationPeriod

depositPeriod :: DepositPeriod

chainBackendOptions :: ChainBackendOptions

data DirectOptions

Constructors

DirectOptions

Fields

networkId :: NetworkId

Network identifier to which we expect to connect.

nodeSocket :: SocketPath

Path to a domain socket used to connect to the server.

data OfflineChainConfig

Constructors

OfflineChainConfig

Fields

offlineHeadSeed :: HeadSeed

Manually provided seed of the offline head.

initialUTxOFile :: FilePath

Path to a json encoded starting UTxO for the offline-mode head.

ledgerGenesisFile :: Maybe FilePath

Path to a shelley genesis file with slot lengths used by the offline-mode chain.

data RunOptions

Constructors

RunOptions

Fields

verbosity :: Verbosity

nodeId :: NodeId

listen :: Host

advertise :: Maybe Host

peers :: [Host]

apiHost :: IP

apiPort :: PortNumber

tlsCertPath :: Maybe FilePath

tlsKeyPath :: Maybe FilePath

monitoringPort :: Maybe PortNumber

hydraSigningKey :: FilePath

hydraVerificationKeys :: [FilePath]

persistenceDir :: FilePath

persistenceRotateAfter :: Maybe (Positive Natural)

chainConfig :: ChainConfig

ledgerConfig :: LedgerConfig

whichEtcd :: WhichEtcd

apiTransactionTimeout :: ApiTransactionTimeout

data ChainConfig

Constructors

Offline OfflineChainConfig

Cardano CardanoChainConfig

data InvalidOptions

Constructors

MaximumNumberOfPartiesExceeded

CardanoAndHydraKeysMismatch

newtype LedgerConfig

Constructors

CardanoLedgerConfig

Fields

cardanoLedgerProtocolParametersFile :: FilePath

newtype GenerateKeyPair

Constructors

GenerateKeyPair

Fields

outputFile :: FilePath

data PublishOptions

Constructors

PublishOptions

Fields

chainBackendOptions :: ChainBackendOptions

publishSigningKey :: FilePath

data BlockfrostChainConfig

Constructors

BlockfrostChainConfig

Fields

projectPath :: FilePath

Path to the blockfrost project file

cardanoSigningKey :: FilePath

Path to the cardano signing key of the internal wallet.

hydraScriptsTxId :: [TxId]

Identifier of transaction holding the hydra scripts to use.

toArgs :: RunOptions -> [String]

Convert an Options instance into the corresponding list of command-line arguments.

This is useful in situations where one wants to programmatically define Options, providing some measure of type safety, without having to juggle with strings.

defaultContestationPeriod :: ContestationPeriod

Default contestation period of 12 hours, aligned with Cardano's safe zone on mainnet. The safe zone is approximately 3 * k / f where k = 2160 (security parameter) and f = 0.05 (active slot coefficient), ensuring finality guarantees. See: https://github.com/cardano-scaling/hydra/issues/2389

defaultDepositPeriod :: DepositPeriod

validateRunOptions :: RunOptions -> Either InvalidOptions ()

Validate cmd line arguments for hydra-node and check if they make sense before actually running the node. Rules we apply: - Check if number of parties is bigger than our hardcoded limit (by looking at loaded hydra or cardano keys and comparing it to the maximumNumberOfParties) - Check that number of loaded hydra keys match with the number of loaded cardano keys (by comparing lengths of the two lists)

genChainPoint :: Gen ChainPoint

commandParser :: Parser Command

runOptionsParser :: Parser RunOptions

Parser for running the cardano-node with all its RunOptions.

publishOptionsParser :: Parser PublishOptions

outputFileParser :: Parser FilePath

defaultPublishOptions :: PublishOptions

Default options as they should also be provided by runOptionsParser.

defaultDirectOptions :: DirectOptions

defaultBlockfrostOptions :: BlockfrostOptions

defaultBFQueryTimeout :: Int

defaultBFRetryTimeout :: Int

chainBackendOptionsParser :: Parser ChainBackendOptions

cardanoSigningKeyFileParser :: Parser FilePath

genFilePath :: String -> Gen FilePath

genDirPath :: Gen FilePath

defaultRunOptions :: RunOptions

Default options as they should also be provided by runOptionsParser.

defaultCardanoChainConfig :: CardanoChainConfig

defaultLedgerConfig :: LedgerConfig

verbosityParser :: Parser Verbosity

nodeIdParser :: Parser NodeId

listenParser :: Parser Host

advertiseParser :: Parser Host

peerParser :: Parser Host

apiHostParser :: Parser IP

apiPortParser :: Parser PortNumber

tlsCertPathParser :: Parser FilePath

tlsKeyPathParser :: Parser FilePath

monitoringPortParser :: Parser PortNumber

hydraSigningKeyFileParser :: Parser FilePath

hydraVerificationKeyFileParser :: Parser FilePath

persistenceDirParser :: Parser FilePath

persistenceRotateAfterParser :: Parser (Positive Natural)

chainConfigParser :: Parser ChainConfig

ledgerConfigParser :: Parser LedgerConfig

whichEtcdParser :: Parser WhichEtcd

apiTransactionTimeoutParser :: Parser ApiTransactionTimeout

cardanoChainConfigParser :: Parser CardanoChainConfig

offlineChainConfigParser :: Parser OfflineChainConfig

networkIdParser :: Parser NetworkId

nodeSocketParser :: Parser SocketPath

blockfrostProjectPathParser :: Parser FilePath

blockfrostQueryTimeoutParser :: Parser Int

blockfrostRetryTimeoutParser :: Parser Int

cardanoLedgerProtocolParametersParser :: Parser FilePath

offlineHeadSeedParser :: Parser HeadSeed

initialUTxOFileParser :: Parser FilePath

ledgerGenesisFileParser :: Parser (Maybe FilePath)

hydraScriptsTxIdsParser :: Parser [TxId]

hydraScriptsTxIdParser :: Parser TxId

hydraScriptsDefaultParser :: Parser [TxId]

cardanoVerificationKeyFileParser :: Parser FilePath

startChainFromParser :: Parser ChainPoint

contestationPeriodParser :: Parser ContestationPeriod

depositPeriodParser :: Parser DepositPeriod

defaultApiTransactionTimeout :: ApiTransactionTimeout

hydraNodeCommand :: ParserInfo Command

parseHydraCommand :: IO Command

Parse command-line arguments into a Option or exit with failure and error message.

parseHydraCommandFromArgs :: [String] -> ParserResult Command

Pure parsing of Option from a list of arguments.

toArgApiPort :: PortNumber -> [String]

toArgNetworkId :: NetworkId -> [String]

toArgNodeSocket :: SocketPath -> [String]

data ParserResult a

Constructors

Success a

Failure (ParserFailure ParserHelp)

CompletionInvoked CompletionResult

renderFailure :: ParserFailure ParserHelp -> String -> (String, ExitCode)

Orphan instances

Arbitrary IP

Instance details

Methods

arbitrary :: Gen IP

shrink :: IP -> [IP]

FromJSON a => FromJSON (Positive a)

Instance details

Methods

parseJSON :: Value -> Parser (Positive a)

parseJSONList :: Value -> Parser [Positive a]

omittedField :: Maybe (Positive a)

ToJSON a => ToJSON (Positive a)

Instance details

Methods

toJSON :: Positive a -> Value

toEncoding :: Positive a -> Encoding

toJSONList :: [Positive a] -> Value

toEncodingList :: [Positive a] -> Encoding

omitField :: Positive a -> Bool