Hydra.Plutus.GoldenSpec

Description

Golden tests of hydra-plutus scripts.

This test suite ensures we do not accidentally change scripts and also persists the plutus scripts of the Hydra protocol as blobs in the repository.

This is also crucial in case we cannot reproduce them exactly as they were originally compiled using plutus-tx; which is not unlikely given we need to have the exact same version of plutus-tx, all its dependencies, and GHC.

Documentation

spec :: Spec

goldenScript :: String -> Script -> Golden Script

Write a golden script on first run and ensure it stays the same on subsequent runs.