Hydra.TUI.Model

Documentation

data TUIEvent tx

TUI-local event wrapper so we can inject async results alongside Hydra node events.

Constructors

NodeEvent (HydraEvent tx)

UTxOQueryResult (Map TxIn (TxOut CtxUTxO))

L1UTxORefresh (Map TxIn (TxOut CtxUTxO))

FuelUTxORefresh (Map TxIn (TxOut CtxUTxO))

TxBuildError Text

data RootState

Constructors

RootState

Fields

nodeHost :: Host

now :: UTCTime

timeZone :: TimeZone

connectedState :: ConnectedState

logState :: LogState

activeTab :: ActiveTab

eventDetailRaw :: Bool

eventHistoryList :: List Name LogMessage

pendingAction :: Maybe Text

l1UTxO :: Maybe (Map TxIn (TxOut CtxUTxO))

fuelVk :: Maybe (VerificationKey PaymentKey)

Verification key of the node's internal wallet, if a fuel key was configured. Used to derive the fuel address and highlight its outputs.

fuelUTxO :: Maybe (Map TxIn (TxOut CtxUTxO))

Last queried UTxO at the fuel address. Display-only; never committed.

previousTab :: ActiveTab

theme :: Theme

recoveryForm :: Maybe (TxIdRadioFieldForm (HydraEvent Tx) Name)

eventHistoryFilter :: EventHistoryFilter

data ConnectedState

Connection to the hydra node.

Constructors

Disconnected

Connected

Fields

connection :: Connection

data IdentifiedState

Constructors

Unidentified

Identified Party

data NetworkState

Connectivity of the hydra node to the hydra network.

Constructors

NetworkConnected

NetworkDisconnected

data ChainSyncedStatus

Synchronization status of the hydra node to the cardano chain.

Constructors

InSync

CatchingUp

data PeerStatus

Constructors

PeerIsConnected

PeerIsDisconnected

PeerIsUnknown

data Connection

Constructors

Connection

Fields

me :: IdentifiedState

peers :: [(Host, PeerStatus)]

networkState :: Maybe NetworkState

chainSyncedStatus :: ChainSyncedStatus

headState :: HeadState

type UTxOCheckboxForm e n = Form (Map TxIn (TxOut CtxUTxO, Bool)) e n

type UTxORadioFieldForm e n = Form (TxIn, TxOut CtxUTxO) e n

type TxIdRadioFieldForm e n = Form TxId e n

type ConfirmingRadioFieldForm e n = Form Bool e n

data OpenScreen

Constructors

OpenHome

LoadingUTxOForIncrement

NoUTxOToIncrement

SelectingUTxO

Fields

selectingUTxOForm :: UTxORadioFieldForm (HydraEvent Tx) Name

SelectingUTxOToDecommit

Fields

selectingUTxOToDecommitForm :: UTxORadioFieldForm (HydraEvent Tx) Name

SelectingUTxOToIncrement

Fields

selectingUTxOToIncrementForm :: UTxORadioFieldForm (HydraEvent Tx) Name

EnteringAmount

Fields

utxoSelected :: (TxIn, TxOut CtxUTxO)

enteringAmountForm :: Form Double (HydraEvent Tx) Name

SelectingRecipient

Fields

utxoSelected :: (TxIn, TxOut CtxUTxO)

amountEntered :: Double

selectingRecipientForm :: Form SelectAddressItem (HydraEvent Tx) Name

EnteringRecipientAddress

Fields

utxoSelected :: (TxIn, TxOut CtxUTxO)

amountEntered :: Double

enteringRecipientAddressForm :: Form AddressInEra (HydraEvent Tx) Name

ConfirmingClose

Fields

confirmingCloseForm :: ConfirmingRadioFieldForm (HydraEvent Tx) Name

data SelectAddressItem

Constructors

ManualEntry

SelectAddress AddressInEra

newtype ClosedState

Constructors

ClosedState

Fields

contestationDeadline :: UTCTime

data HeadState

Constructors

Idle

Active

Fields

activeLink :: ActiveLink

data PendingIncrementStatus

Constructors

PendingDeposit

FinalizingDeposit

data PendingIncrement

Constructors

PendingIncrement

Fields

utxoToCommit :: UTxO

deposit :: TxId

depositDeadline :: UTCTime

status :: PendingIncrementStatus

data ActiveLink

Constructors

ActiveLink

Fields

utxo :: UTxO

pendingUTxOToDecommit :: UTxO

pendingIncrements :: [PendingIncrement]

parties :: [Party]

headId :: HeadId

activeHeadState :: ActiveHeadState

data ActiveHeadState

Constructors

Open

Fields

openState :: OpenScreen

Closed

Fields

closedState :: ClosedState

FanoutPossible

Final

type Name = Text

data ActiveTab

Constructors

MainTab

FundsTab

EventHistoryTab

ModalTab

selectingUTxOToIncrementFormL :: Traversal' OpenScreen (UTxORadioFieldForm (HydraEvent Tx) Name)

selectingUTxOToDecommitFormL :: Traversal' OpenScreen (UTxORadioFieldForm (HydraEvent Tx) Name)

selectingUTxOFormL :: Traversal' OpenScreen (UTxORadioFieldForm (HydraEvent Tx) Name)

selectingRecipientFormL :: Traversal' OpenScreen (Form SelectAddressItem (HydraEvent Tx) Name)

enteringRecipientAddressFormL :: Traversal' OpenScreen (Form AddressInEra (HydraEvent Tx) Name)

enteringAmountFormL :: Traversal' OpenScreen (Form Double (HydraEvent Tx) Name)

confirmingCloseFormL :: Traversal' OpenScreen (ConfirmingRadioFieldForm (HydraEvent Tx) Name)

openStateL :: Traversal' ActiveHeadState OpenScreen

closedStateL :: Traversal' ActiveHeadState ClosedState

timeZoneL :: Lens' RootState TimeZone

themeL :: Lens' RootState Theme

recoveryFormL :: Lens' RootState (Maybe (TxIdRadioFieldForm (HydraEvent Tx) Name))

previousTabL :: Lens' RootState ActiveTab

pendingActionL :: Lens' RootState (Maybe Text)

nowL :: Lens' RootState UTCTime

nodeHostL :: Lens' RootState Host

logStateL :: Lens' RootState LogState

l1UTxOL :: Lens' RootState (Maybe (Map TxIn (TxOut CtxUTxO)))

fuelVkL :: Lens' RootState (Maybe (VerificationKey PaymentKey))

fuelUTxOL :: Lens' RootState (Maybe (Map TxIn (TxOut CtxUTxO)))

eventHistoryListL :: Lens' RootState (List Name LogMessage)

eventHistoryFilterL :: Lens' RootState EventHistoryFilter

eventDetailRawL :: Lens' RootState Bool

connectedStateL :: Lens' RootState ConnectedState

activeTabL :: Lens' RootState ActiveTab

connectionL :: Traversal' ConnectedState Connection

peersL :: Lens' Connection [(Host, PeerStatus)]

networkStateL :: Lens' Connection (Maybe NetworkState)

meL :: Lens' Connection IdentifiedState

headStateL :: Lens' Connection HeadState

chainSyncedStatusL :: Lens' Connection ChainSyncedStatus

activeLinkL :: Traversal' HeadState ActiveLink

utxoL :: Lens' ActiveLink UTxO

pendingUTxOToDecommitL :: Lens' ActiveLink UTxO

pendingIncrementsL :: Lens' ActiveLink [PendingIncrement]

partiesL :: Lens' ActiveLink [Party]

headIdL :: Lens' ActiveLink HeadId

activeHeadStateL :: Lens' ActiveLink ActiveHeadState

eventHistoryListName :: Name

mainUTxOViewportName :: Name

fundsL2ViewportName :: Name

fundsL1ViewportName :: Name

fundsFuelViewportName :: Name

emptyEventHistoryList :: List Name LogMessage

emptyConnection :: Connection

newActiveLink :: [Party] -> HeadId -> ActiveLink

isModalOpen :: RootState -> Bool

recoverHeadState :: UTCTime -> HeadState -> NodeState Tx -> HeadState