Hydra.Tx.Close

Construction

type PointInTime = (SlotNo, UTCTime)

data OpenThreadOutput

Representation of the Head output after a CollectCom transaction.

Constructors

OpenThreadOutput

Fields

openThreadUTxO :: (TxIn, TxOut CtxUTxO)

openContestationPeriod :: ContestationPeriod

openParties :: [Party]

closeTx

Arguments

:: ScriptRegistry
Published Hydra scripts to reference.

-> VerificationKey PaymentKey
Party who's authorizing this transaction

-> HeadId
Head identifier

-> SnapshotVersion
Last known version of the open head.

-> ConfirmedSnapshot Tx
Snapshot with instructions how to close the head.

-> SlotNo
Lower validity slot number, usually a current or quite recent slot number.

-> PointInTime
Upper validity slot and UTC time to compute the contestation deadline time.

-> OpenThreadOutput
Everything needed to spend the Head state-machine output.

-> IncrementalAction

-> Tx

Create a transaction closing a head with either the initial snapshot or with a multi-signed confirmed snapshot.

Observation

data CloseObservation

Constructors

CloseObservation

Fields

headId :: HeadId

snapshotNumber :: SnapshotNumber

contestationDeadline :: UTCTime

observeCloseTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe CloseObservation

Identify a close tx by lookup up the input spending the Head output and decoding its redeemer.