Hydra.Tx.Contract.Close.CloseInitial

Documentation

data CloseInitialMutation

Constructors

MutateCloseContestationDeadline'

healthyCloseSnapshotVersion :: SnapshotVersion

healthyCloseInitialTx :: (Tx, UTxO)

Healthy close transaction for the specific case were we close a head with the initial UtxO, that is, no snapshot have been agreed upon and signed by the head members yet.

healthyInitialOpenDatum :: State

genCloseInitialMutation :: (Tx, UTxO) -> Gen SomeMutation

genMutatedDeadline :: Gen POSIXTime

Generate not acceptable, but interesting deadlines.