Hydra.Tx.Contract.Close.CloseUsed

Documentation

healthyOutdatedSnapshotNumber :: SnapshotNumber

healthyOutdatedSnapshotVersion :: SnapshotVersion

healthyOutdatedSnapshot :: Snapshot Tx

Decommit snapshot which refers to the previous open state version. That is, healthyOpenStateVersion - 1.

healthyOutdatedOpenDatum :: State

healthyOpenStateVersion :: SnapshotVersion

In the outdated case, the used snapshot version is exactly one lower than the open state version.

healthyOutdatedConfirmedClosingSnapshot :: ConfirmedSnapshot Tx

healthyCloseOutdatedTx :: (Tx, UTxO)

data CloseMutation

Constructors

NotContinueContract
Ensures collectCom does not allow any output address but νHead.

MutateSignatureButNotSnapshotNumber
Ensures the snapshot signature is multisigned by all valid Head participants.

Invalidates the tx by changing the redeemer signature but not the snapshot number in output head datum.

MutateSnapshotNumberButNotSignature
Ensures the snapshot number is consistent with the signature.

Invalidates the tx by changing the snapshot number in resulting head output but not the redeemer signature.

MutateSnapshotVersion
Check the snapshot version is preserved from last open state.

SnapshotNotSignedByAllParties
Ensures the close snapshot is multisigned by all Head participants by changing the parties in the input head datum. If they do not align the multisignature will not be valid anymore.

MutateRequiredSigner
Ensures close is authenticated by a one of the Head members by changing the signer used on the tx to not be one of PTs.

MutateNoRequiredSigner
Ensures close is authenticated by a one of the Head members by changing the signer used on the tx to be empty.

MutateMultipleRequiredSigner
Ensures close is authenticated by a one of the Head members by changing the signer used on the tx to have multiple signers (including the signer to not fail for SignerIsNotAParticipant).

MutateCloseUTxOHash
Invalidates the tx by changing the utxo hash in resulting head output.

Ensures the output state is consistent with the redeemer.

MutateCloseUTxOToDecommitHash
Invalidates the tx by changing the utxo to decommit hash in resulting head output.

Ensures the output state is consistent with the redeemer.

MutateCloseType
Invalidates the tx by changing claimed closing type. i.e. claim the snapshot is current but provide signatures from an previous version

MutateCloseSignatures
Invalidates the tx by changing the signatures in redeemer.

Ensures the output state is consistent with the redeemer.

MutatePartiesInOutput
Ensures parties do not change between head input datum and head output datum.

MutateHeadIdInOutput
Ensures headId do not change between head input datum and head output datum.

MutateInfiniteLowerBound
Invalidates the tx by changing the lower bound to be non finite.

MutateInfiniteUpperBound
Invalidates the tx by changing the upper bound to be non finite.

MutateContestationDeadline
Invalidates the tx by changing the contestation deadline to not satisfy `contestationDeadline = upperBound + contestationPeriod`.

MutateValidityInterval
Invalidates the tx by changing the lower and upper bound to be not bounded as per spec `upperBound - lowerBound <= contestationPeriod`.

This also changes the resulting `head output` contestation deadline to be valid, so it satisfy `contestationDeadline = upperBound + contestationPeriod`.

CloseFromDifferentHead
Ensure the Head cannot be closed with correct authentication from a different Head. We simulate this by changing the head policy id of the ST and PTs to be of a different head - a real attack would be to add inputs with those tokens on top of spending the head output, a bit like a double satisfaction attack. Note that the token name stays the same and consistent with the signer. This will cause authentication failure because the signer's PT, although with a consistent name, is not from the right head (has a different policy id than in the datum).

MutateTokenMintingOrBurning
Minting or burning of tokens should not be possible in close.

MutateContesters
Invalidates the tx by changing the contesters to be non empty.

MutateValueInOutput
Invalidates the tx by changing output values arbitrarily to be different (not preserved) from the head.

Ensures values are preserved between head input and output.

MutateContestationPeriod
Invalidate the tx by changing the contestation period.

genCloseOutdatedMutation :: (Tx, UTxO) -> Gen SomeMutation

genMutatedDeadline :: Gen POSIXTime

Generate not acceptable, but interesting deadlines.