Hydra.Tx.Contract.Contest.Healthy

Documentation

healthyContestTx :: (Tx, UTxO)

Healthy contest tx where the contester is the first one to contest and correctly pushing out the deadline by the contestation period.

healthyContestSnapshotNumber :: SnapshotNumber

healthyCloseSnapshotVersion :: SnapshotVersion

healthyClosedUTxO :: UTxO

healthyContestUTxO :: UTxO

splitContestUTxO :: (UTxO, UTxO)

splitUTxOInHead :: UTxO

splitUTxOToDecommit :: UTxO

healthyContestSnapshot :: Snapshot Tx

healthyClosedState :: State

healthyContestUTxOHash :: BuiltinByteString

healthyContestUTxOToDecommitHash :: BuiltinByteString

healthyClosedUTxOHash :: BuiltinByteString

healthyClosedSnapshotNumber :: SnapshotNumber

healthySlotNo :: SlotNo

healthyClosedHeadTxIn :: TxIn

healthyClosedHeadTxOut :: TxOut CtxUTxO

healthyOnChainContestationPeriod :: ContestationPeriod

healthyContestationPeriod :: ContestationPeriod

healthyContestationPeriodSeconds :: Integer

healthyParticipants :: [VerificationKey PaymentKey]

healthyContesterVerificationKey :: VerificationKey PaymentKey

healthySigningKeys :: [SigningKey HydraKey]

healthyParties :: [Party]

healthyOnChainParties :: [Party]

healthySignature :: SnapshotNumber -> MultiSignature (Snapshot Tx)

healthyContestationDeadline :: UTCTime