Hydra.Tx.Contract.Decrement

Documentation

healthyDecrementTx :: (Tx, UTxO)

somePartyCardanoVerificationKey :: VerificationKey PaymentKey

healthySigningKeys :: [SigningKey HydraKey]

healthyParticipants :: [VerificationKey PaymentKey]

healthyParties :: [Party]

healthyOnChainParties :: [Party]

healthySignature :: MultiSignature (Snapshot Tx)

healthySnapshotNumber :: SnapshotNumber

healthySnapshotVersion :: SnapshotVersion

healthySnapshot :: Snapshot Tx

splitDecommitUTxO :: UTxO -> (UTxO, UTxO)

healthyContestationPeriod :: ContestationPeriod

healthyUTxO :: UTxO

healthyDatum :: State

data DecrementMutation

Constructors

ChangePartiesInOuput
Ensures parties do not change between head input datum and head output datum.

ProduceInvalidSignatures
Produce invalid signature by changing signers in the redeemer

AlterRequiredSigner
Ensures decrement is authenticated by one of the Head members by changing the signer used on the tx to not be one of PTs.

ChangeDecrementedValue
Mutate the output value to produce different UTxO' hash to the one in the signed Snapshot.

ChangeHeadValue
Invalidates the tx by changing the output values arbitrarily to be different (not preserved) from the head.

Ensures values are preserved between head input and output.

DropDecommitOutput
Drop one of the decommit outputs from the tx. This should trigger snapshot signature validation to fail.

ExtractSomeValue

UseDifferentSnapshotVersion
Invalidates the tx by changing the snapshot version in resulting head output.

genDecrementMutation :: (Tx, UTxO) -> Gen SomeMutation