Hydra.Tx.Contract.Deposit

Description

Healthy deposit transactions and mutations As no Hydra script is run in these transactions, the mutations here should make the deposit transaction not observed as a valid deposi.

Documentation

genHealthyDepositTx :: Gen (Tx, UTxO)

healthyDeadline :: UTCTime

healthyDeadlineSlot :: SlotNo

data DepositMutation

Constructors

MutateDepositOutputValue
Change the output value to a subset of the deposited value. This simulates an attack where someone claims to have deposited more than they actually did.

genDepositMutation :: (Tx, UTxO) -> Gen SomeMutation