Hydra.Tx.Contract.Recover

Documentation

healthyRecoverTx :: (Tx, UTxO)

recoverSlotNo :: SlotNo

depositDeadline :: UTCTime

depositTransaction :: Tx

healthyDepositUTxO :: UTxO

headCS :: CurrencySymbol

headPolicyId :: PolicyId

depositScriptUTxO :: UTxO

depositTxIn :: TxIn

data RecoverMutation

Constructors

MutateDepositDeadline
Move the deposit deadline further so that the recover lower bound is not after the deadline

MutateRecoverOutput
Change the recover output so that the datum commit hash does not match

RemoveTxValidityLowerBound
Remove the lower bound from the recover transaction

genRecoverMutation :: (Tx, UTxO) -> Gen SomeMutation