Hydra.Tx.Crypto

Description

Hydra multi-signature credentials and cryptographic primitives used to sign and verify snapshots (or any messages) within the Hydra protocol.

We are re-using the Key interface of 'cardano-api' for a consistent representation. For example: Cardano credentials are 'VerificationKey PaymentKey', Hydra credentials are 'VerificationKey HydraKey'.

Currently MultiSignature interface is only supporting naiive, concatenated multi-signatures and will change when we adopt aggregated multi-signatures including aggregate keys.

Cardano Key interface

class (Eq (VerificationKey keyrole), Show (VerificationKey keyrole), SerialiseAsRawBytes (Hash keyrole), HasTextEnvelope (VerificationKey keyrole), HasTextEnvelope (SigningKey keyrole)) => Key keyrole where

Associated Types

data VerificationKey keyrole

data SigningKey keyrole

Methods

getVerificationKey :: SigningKey keyrole -> VerificationKey keyrole

deterministicSigningKey :: AsType keyrole -> Seed -> SigningKey keyrole

deterministicSigningKeySeedSize :: AsType keyrole -> Word

verificationKeyHash :: VerificationKey keyrole -> Hash keyrole

Hydra specifics

data family Hash keyrole

data family AsType t

data family SigningKey keyrole

data family VerificationKey keyrole

newtype MultiSignature a

Naiively aggregated multi-signatures.

Constructors

HydraMultiSignature

Fields

multiSignature :: [Signature a]

data HydraKey

Hydra keys (keyrole) which can be used to sign and verify messages, as well as aggregate multi-signatures.

newtype Signature a

Signature of a, not containing the actual payload.

Constructors

HydraSignature (SigDSIGN Ed25519DSIGN)

data Verified

A result type for multisigs verification providing some information in case of failure.

This type is of course structurally equivalent to `Maybe [VerificationKey HydraKey]` but it's much more explicit.

Constructors

Verified

FailedKeys

Fields

failedKeys :: [VerificationKey HydraKey]

KeyNumberMismatch

pattern HydraKeyHash :: Hash Blake2b_256 (VerificationKey HydraKey) -> Hash HydraKey

pattern AsHydraKey :: AsType HydraKey

pattern HydraSigningKey :: SignKeyDSIGN Ed25519DSIGN -> SigningKey HydraKey

pattern HydraVerificationKey :: VerKeyDSIGN Ed25519DSIGN -> VerificationKey HydraKey

sign :: SignableRepresentation a => SigningKey HydraKey -> a -> Signature a

Sign some value a with the provided SigningKey.

toPlutusSignatures :: MultiSignature a -> [Signature]

aggregate :: [Signature a] -> MultiSignature a

Combine multiple signatures of a into a 'MultiSignature a'.

generateSigningKey :: ByteString -> SigningKey HydraKey

Create a new SigningKey from a ByteString seed. The created keys are not random and insecure, so don't use this in production code!

verify :: SignableRepresentation a => VerificationKey HydraKey -> Signature a -> a -> Bool

Verify a given 'Signature a' and value a using provided VerificationKey.

deserialiseFromRawBytesBase16 :: SerialiseAsRawBytes a => ByteString -> Either String (UsingRawBytesHex a)

aggregateInOrder :: Ord k => Map k (Signature a) -> [k] -> MultiSignature a

Like aggregate, but use order of given list of keys instead. FIXME(AB): This function exists solely because the order of signatures matters on-chain, and it should match the order of parties as declared in the initTx. This should disappear once we use a proper multisignature scheme

verifyMultiSignature :: SignableRepresentation a => [VerificationKey HydraKey] -> MultiSignature a -> a -> Verified

Verify a given 'MultiSignature a' and value a provided a list of VerificationKey.

Note that order of keys is relevant and that length of signature and multisignature list needs to be the same.

Orphan instances

IsString TxId

Instance details

Methods

fromString :: String -> TxId

IsString (Hash BlockHeader)

Instance details

Methods

fromString :: String -> Hash BlockHeader

IsString (Hash ScriptData)

Instance details

Methods

fromString :: String -> Hash ScriptData

SerialiseAsRawBytes a => IsString (UsingRawBytesHex a)

Instance details

Methods

fromString :: String -> UsingRawBytesHex a