Hydra.Tx.HeadId

HeadId

newtype HeadId

Uniquely identifies a Hydra Head.

Constructors

UnsafeHeadId ByteString

currencySymbolToHeadId :: MonadFail m => CurrencySymbol -> m HeadId

headIdToPolicyId :: MonadFail m => HeadId -> m PolicyId

headIdToCurrencySymbol :: HeadId -> CurrencySymbol

mkHeadId :: PolicyId -> HeadId

HeadSeed

newtype HeadSeed

Unique seed to create a HeadId

XXX: This might actually be the HeadId to the protocol and users? Then the policy id of the cardano-specific implementation (being the result of minting policy + seed) stays internal. A drawback is, that the seed is not such a good "key" to find things about this head on explorers and indexers.

Constructors

UnsafeHeadSeed ByteString

headSeedToTxIn :: MonadFail m => HeadSeed -> m TxIn

txInToHeadSeed :: TxIn -> HeadSeed