Hydra.Tx.ScriptRegistry

Description

A data-type to keep track of reference Hydra scripts published on-chain, and needed to construct transactions leveraging reference inputs.

Documentation

data ScriptRegistry

Hydra scripts published as reference scripts at these UTxO.

Constructors

ScriptRegistry

Fields

initialReference :: (TxIn, TxOut CtxUTxO)

commitReference :: (TxIn, TxOut CtxUTxO)

headReference :: (TxIn, TxOut CtxUTxO)

data NewScriptRegistryException

Constructors

MissingScript

Fields

scriptName :: Text

scriptHash :: ScriptHash

discoveredScripts :: Set ScriptHash

newScriptRegistry :: UTxO -> Either NewScriptRegistryException ScriptRegistry

Create a script registry from a UTxO containing outputs with reference scripts. This will return Nothing if one or all of the references could not be found.

registryUTxO :: ScriptRegistry -> UTxO

Get the UTxO that corresponds to a script registry.

*Property**:

newScriptRegistry (registryUTxO r) === Just r