Hydra.Tx.Snapshot

SnapshotNumber and SnapshotVersion

newtype SnapshotNumber

Constructors

UnsafeSnapshotNumber Natural

fromChainSnapshotNumber :: SnapshotNumber -> SnapshotNumber

newtype SnapshotVersion

Constructors

UnsafeSnapshotVersion Natural

fromChainSnapshotVersion :: SnapshotVersion -> SnapshotVersion

Snapshot

data Snapshot tx

Constructors

Snapshot

Fields

headId :: HeadId

version :: SnapshotVersion

Open state version this snapshot is based on. Spec: v

number :: SnapshotNumber

Monotonically increasing snapshot number. Spec: s

confirmed :: [tx]

The set of transactions that lead to $sel:utxo:Snapshot. Spec: T

utxo :: UTxOType tx

Snaspshotted UTxO set. Spec: U

utxoToCommit :: Maybe (UTxOType tx)

UTxO to be committed. Spec: Uα

utxoToDecommit :: Maybe (UTxOType tx)

UTxO to be decommitted. Spec: Uω

ConfirmedSnapshot

data ConfirmedSnapshot tx

A snapshot that can be used to close a head with. Either the initial one, or when it was signed by all parties, i.e. it is confirmed.

Constructors

InitialSnapshot

Fields

headId :: HeadId

initialUTxO :: UTxOType tx

ConfirmedSnapshot

Fields

snapshot :: Snapshot tx

signatures :: MultiSignature (Snapshot tx)

getSnapshot :: ConfirmedSnapshot tx -> Snapshot tx

Safely get a Snapshot from a confirmed snapshot.

NOTE: While we could use $sel:snapshot:InitialSnapshot directly, this is a record-field accessor which may become partial (and lead to unnoticed runtime errors) if we ever add a new branch to the sumtype. So, we explicitly define a getter which will force us into thinking about changing the signature properly if this happens.