Hydra.Version

Description

Provides version numbers from calling git on build time or from an embedded string.

The former is based on the gitrev package with a Maybe interface around it, while the embedding is done using a special c-array placeholder in cbits/revision.c

Documentation

unknownVersion :: String

Identifier to be used when no revision can be found.

This is also the default used in gitrev.

gitDescribe :: Maybe String

Determine the version on build time using `git describe`. FIXME: This does not change when hydra-prelude is not re-compiled

gitRevision :: Maybe String

Determine the version on build time using `git rev-parse`. FIXME: This does not change when hydra-prelude is not re-compiled

c_gitrev :: CString

embeddedRevision :: Maybe String

The git revision embedded at a special place holder in the object/binary. NOTE: Keep this consistent with what is hard-coded in 'cbits/revision.c'