Test.Hydra.Fixture

Description

Test and example values used across hydra-node tests.

Documentation

alice :: Party

Our beloved alice, bob, and carol.

bob :: Party

Our beloved alice, bob, and carol.

carol :: Party

Our beloved alice, bob, and carol.

aliceSk :: SigningKey HydraKey

Hydra signing keys for alice, bob, and carol.

bobSk :: SigningKey HydraKey

Hydra signing keys for alice, bob, and carol.

carolSk :: SigningKey HydraKey

Hydra signing keys for alice, bob, and carol.

aliceVk :: VerificationKey HydraKey

Hydra verification keys for alice, bob, and carol.

bobVk :: VerificationKey HydraKey

Hydra verification keys for alice, bob, and carol.

carolVk :: VerificationKey HydraKey

Hydra verification keys for alice, bob, and carol.

alicePVk :: VerificationKey PaymentKey

Cardano payment keys for alice, bob, and carol.

bobPVk :: VerificationKey PaymentKey

Cardano payment keys for alice, bob, and carol.

carolPVk :: VerificationKey PaymentKey

Cardano payment keys for alice, bob, and carol.

cperiod :: ContestationPeriod

testHeadId :: HeadId

testHeadSeed :: HeadSeed

deriveOnChainId :: Party -> OnChainId

Derive some OnChainId from a Hydra party. In the real protocol this is currently not done, but in this simulated chain setting this is definitely fine.

genForParty :: Gen a -> Party -> a

Generate some a given the Party as a seed. NOTE: While this is useful to generate party-specific values, it DOES depend on the generator used. For example, `genForParty genVerificationKey` and `genForParty (fst $ genKeyPair)` do not yield the same verification keys!

testEnvironment :: Environment

An environment fixture for testing.

testHeadParameters :: HeadParameters

Head parameters fixture for testing.