Test.Util

Documentation

noopCallback :: Applicative m => NetworkCallback msg m

shouldRunInSim :: (forall s. IOSim s a) -> IO a

Run given action in IOSim and rethrow any exceptions.

printTrace :: forall log a. (Typeable log, ToJSON log) => Proxy log -> SimTrace a -> Text

Utility function to dump logs given a SimTrace.

shouldBe :: (HasCallStack, MonadThrow m, Eq a, Show a) => a -> a -> m ()

Lifted variant of Hspec's shouldBe.

shouldNotBe :: (HasCallStack, MonadThrow m, Eq a, Show a) => a -> a -> m ()

Lifted variant of Hspec's shouldNotBe.

shouldSatisfy :: (HasCallStack, MonadThrow m, Show a) => a -> (a -> Bool) -> m ()

Lifted variant of Hspec's shouldSatisfy.

traceInIOSim :: Typeable a => Tracer (IOSim s) a

A Tracer that works in IOSim monad. This tracer uses the Output event which uses converts value traced to Dynamic which requires Typeable constraint. To retrieve the trace use selectTraceEventsDynamic applied to the correct type.

traceDebug :: (Applicative m, ToJSON a) => Tracer m a

Useful when one needs to also trace logs to stderr. Thanks to the monoidal nature of Tracer it's straightforward to add this to any existing tracer:

@ someCode tracer = do foo <- makeFoo withTracer (tr <> traceDebug) SomeTraceFoo ... @

propCollisionResistant :: (Show a, Eq a) => String -> Gen a -> Spec

This creates an hspec test case about a property which ensures the given generator does not produce equals values within a reasonable number of generated values.

isContinuous :: (Eq a, Enum a) => [a] -> Bool

Predicate which decides whether given list is continuous.

isStrictlyMonotonic :: Ord a => [a] -> Bool

Predicate which decides whether given list is strictly monotonic.

waitEq :: (HasCallStack, Eq a, Show a) => IO a -> NominalDiffTime -> a -> IO ()

Wait up to some time for a function to yield an equal value.

waitMatch :: (HasCallStack, Show a) => IO a -> NominalDiffTime -> (a -> Maybe b) -> IO b

Wait up to some time for a function to return a value that satisfies given predicate.

captureTracer :: Text -> IO (Tracer IO a, IO [Envelope a])

Create a tracer that captures all messages and a function to retrieve all traces captured. XXX: This is duplicated in MithrilSpec in hydra-cluster, but can't (easily) be moved to the Test Prelude because of the dependency on Hydra.Logging.