Hydra.Chain.Direct.TxSpec

Description

Remainder of tests covering observation and tx creation by the "direct" chain component. XXX: This does not have a corresponding "source" module which it tests.

Documentation

spec :: Spec

propHasValidAuxData :: Tx -> Property

Check auxiliary data of a transaction against pparams and whether the aux data hash is consistent.

propIsSubsetOf :: (Show a, Ord a) => Set a -> Set a -> Property

Check whether one set isSubsetOf of another with nice counter examples.

propIsSubmapOf :: (Show k, Show v, Ord k, Eq v) => Map k v -> Map k v -> Property

Check whether one map isSubmapOf of another with nice counter examples.

genBlueprintTxWithUTxO :: Gen (UTxO, Tx)

genMetadata :: Gen TxMetadataInEra

getAuxMetadata :: Tx -> Map Word64 Metadatum

prop_interestingBlueprintTx :: Property