Hydra.Chain.Direct.TxSpec

Description

Unit tests for our "hand-rolled" transactions as they are used in the "direct" chain component.

Documentation

spec :: Spec

propHasValidAuxData :: Tx -> Property

Check auxiliary data of a transaction against pparams and whether the aux data hash is consistent.

propIsSubsetOf :: (Show a, Ord a) => Set a -> Set a -> Property

Check whether one set isSubsetOf of another with nice counter examples.

propIsSubmapOf :: (Show k, Show v, Ord k, Eq v) => Map k v -> Map k v -> Property

Check whether one map isSubmapOf of another with nice counter examples.

genBlueprintTxWithUTxO :: Gen (UTxO, Tx)

genMetadata :: Gen TxMetadataInEra

getAuxMetadata :: Tx -> Map Word64 Metadatum

prop_interestingBlueprintTx :: Property

withinTxExecutionBudget :: EvaluationReport -> Property

generateCommitUTxOs :: [Party] -> Gen (Map TxIn (TxOut CtxUTxO, UTxO))

Generate a UTXO representing commit outputs for a given list of Party. NOTE: Uses testPolicyId for the datum. NOTE: We don't generate empty commits and it is used only at one place so perhaps move it? FIXME: This function is very complicated and it's hard to understand it after a while

prettyEvaluationReport :: EvaluationReport -> String

genAbortableOutputs :: [Party] -> Gen ([(TxIn, TxOut CtxUTxO)], [(TxIn, TxOut CtxUTxO, UTxO)])

assetNameFromVerificationKey :: VerificationKey PaymentKey -> AssetName

fst3 :: (a, b, c) -> a

third :: (a, b, c) -> c

drop2nd :: (a, b, c) -> (a, c)

drop3rd :: (a, b, c) -> (a, b)

tripleToPair :: (a, b, c) -> (a, (b, c))