Hydra.Chain.Direct.TxTraceSpec

Description

Stateful model-based testing of the transactions created by the Direct chain modules.

The model is focusing on transitions between Open and Closed states of the head.

Given an initial UTxO, the model generates a plausible sequence of snapshots that an honest node would approve. That is, the total balance of UTxO remains constant and utxoToDecommit is only allowed to clear if the version is incremented. Consequently, also all snapshots are correctly signed (we don't test handling of adverarial signatures). UTxOs are simplified such that they are A-E items, where each maps to an arbitrary real UTxO.

From this sequence of valid snapshots, possible Decrement and Close actions are generated, along with occasional Contest and consequential Fanout.

Documentation

spec :: Spec

coversInterestingActions :: Testable p => Actions Model -> p -> Property

prop_runActions :: Actions Model -> Property

============================== MODEL WORLD ==========================

data SingleUTxO

Constructors

A

B

C

D

E

F

G

H

I

type ModelUTxO = [SingleUTxO]

data Model

Constructors

Model

Fields

headState :: State

knownSnapshots :: [ModelSnapshot]

List of off-chain snapshots, from most recent to oldest.

currentVersion :: SnapshotVersion

currentSnapshotNumber :: SnapshotNumber

closedSnapshotNumber :: SnapshotNumber

alreadyContested :: [Actor]

utxoInHead :: ModelUTxO

pendingDeposit :: ModelUTxO

pendingDecommit :: ModelUTxO

latestSnapshotNumber :: [ModelSnapshot] -> SnapshotNumber

data ModelSnapshot

Model of a real snapshot which contains a SnapshotNumber but also our simplified form of UTxO.

Constructors

ModelSnapshot

Fields

version :: SnapshotVersion

number :: SnapshotNumber

inHead :: ModelUTxO

toCommit :: ModelUTxO

toDecommit :: ModelUTxO

data State

Constructors

Open

Closed

Final

data Actor

Constructors

Alice

Bob

Carol

data TxResult

Result of constructing and performing a transaction. Notably there are three stages to this which can fail: construction, validation, and observation. Results from all stages are needed to express post-conditions.

Constructors

TxResult

Fields

constructedTx :: Either String Tx

spendableUTxO :: UTxO

validationError :: Maybe String

observation :: HeadObservation

============================== REAL WORLD ==========================

newtype AppM a

Application monad to perform model actions. Currently it only keeps a UTxO which is updated whenever transactions are valid in performTx.

Constructors

AppM

Fields

runAppM :: ReaderT (IORef (Maybe TxId, UTxO)) IO a

performTx :: Show err => Action Model a -> Either err Tx -> AppM TxResult

Perform a transaction by evaluating and observing it. This updates the UTxO in the AppM if a transaction is valid and produces a TxResult that can be used to assert expected success / failure.

getValidationError :: Tx -> UTxO -> Maybe String

Fixtures and glue code

allActors :: [Actor]

List of all model actors corresponding to the fixtures used.

realWorldModelUTxO :: ModelUTxO -> UTxO

Map a ModelUTxO to a real-world UTxO.

signedSnapshot :: ModelSnapshot -> (Snapshot Tx, MultiSignature (Snapshot Tx))

A correctly signed snapshot. Given a snapshot number a snapshot signed by all participants (alice, bob and carol) with some UTxO contained is produced.

confirmedSnapshot :: ModelSnapshot -> ConfirmedSnapshot Tx

A confirmed snapshot (either initial or later confirmed), based onTxTra signedSnapshot.

openHeadUTxO :: UTxO

UTxO of the open head on-chain. NOTE: This uses fixtures for headId, parties, and cperiod.

newDepositTx :: Action Model a -> ModelUTxO -> AppM (Either String Tx)

Creates a deposit transaction using given UTxO.

newIncrementTx :: Actor -> ConfirmedSnapshot Tx -> AppM (Either IncrementTxError Tx)

Creates a increment transaction using given utxo and given snapshot.

newDecrementTx :: Actor -> ConfirmedSnapshot Tx -> AppM (Either DecrementTxError Tx)

Creates a decrement transaction using given utxo and given snapshot.

newCloseTx :: Actor -> SnapshotVersion -> ConfirmedSnapshot Tx -> AppM (Either CloseTxError Tx)

Creates a transaction that closes openHeadUTxO with given the snapshot. NOTE: This uses fixtures for headId, parties (alice, bob, carol), contestation period and also claims to close at time 0 resulting in a contestation deadline of 0 + cperiod.

newContestTx :: Actor -> SnapshotVersion -> ConfirmedSnapshot Tx -> AppM (Either ContestTxError Tx)

Creates a contest transaction using given utxo and contesting with given snapshot. NOTE: This uses fixtures for headId, contestation period and also claims to contest at time 0.

newFanoutTx :: Actor -> ModelUTxO -> ModelUTxO -> ModelUTxO -> AppM (Either FanoutTxError Tx)

Creates a fanout transaction using given utxo. NOTE: This uses fixtures for seedTxIn and contestation period. Consequently, the lower bound used is precisely at the maximum deadline slot as if everyone contested.

alicePVk :: VerificationKey PaymentKey

Cardano payment keys for alice, bob, and carol.

bobPVk :: VerificationKey PaymentKey

Cardano payment keys for alice, bob, and carol.

carolPVk :: VerificationKey PaymentKey

Cardano payment keys for alice, bob, and carol.

actorChainContext :: Actor -> ChainContext

Fixture for the chain context of a model Actor on testNetworkId. Uses a generated ScriptRegistry.

testScriptRegistry :: ScriptRegistry

Helpers

runPostconditionM' :: Monad m => PostconditionM' m () -> PostconditionM m Bool

Run a short-cutting variant of PostconditionM which produces True if it reaches the end, or False if fail is used.

newtype PostconditionM' m a

Constructors

PostconditionM' (ExceptT (Maybe String) (PostconditionM m) a)

fulfilled :: Monad m => PostconditionM' m ()

Short-cut a postcondition monad like fail, but in a successful way. This is useful to not have unrelated counterexample' outputs.

counterexample' :: Monad m => String -> PostconditionM' m ()

Add given message in case the postcondition fails.

expectValid :: Monad m => TxResult -> (HeadObservation -> PostconditionM' m a) -> PostconditionM' m a

Assertion helper to check whether a TxResult was valid and the expected HeadObservation could be made. To be used in postcondition.

expectInvalid :: Monad m => TxResult -> PostconditionM' m ()

Assertion helper to check whether a TxResult was invalid or construction failed.