Hydra.HeadLogicSpec

Description

Unit tests of the protocol logic in HeadLogic. These are very fine grained and specific to individual steps in the protocol. More high-level of the protocol logic, especially between multiple parties can be found in BehaviorSpec.

Documentation

spec :: Spec

Properties

prop_ignoresUnrelatedOnInitTx :: Property

genClosedState :: Gen (NodeState SimpleTx)

Utilities

receiveMessage :: Message tx -> Input tx

Create a network input about a received protocol message with default ttl and alice as the sender.

receiveMessageFrom :: Party -> Message tx -> Input tx

Create a network input about a received protocol message with default ttl from given sender.

chainEffect :: PostChainTx SimpleTx -> Effect SimpleTx

Create a chain effect with fixed chain state and slot.

observeTxAtSlot :: Natural -> OnChainTx SimpleTx -> Input SimpleTx

Create an observation chain input with chain state at given slot.

observeTx :: OnChainTx SimpleTx -> Input SimpleTx

Create an observation chain input with fixed chain state and slot.

connectivityChanged :: TTL -> Connectivity -> Input SimpleTx

inIdleState :: NodeState SimpleTx

inUnsyncedIdleState :: NodeState SimpleTx

inInitialState :: [Party] -> NodeState SimpleTx

inOpenState :: [Party] -> NodeState SimpleTx

inOpenState' :: [Party] -> CoordinatedHeadState SimpleTx -> NodeState SimpleTx

inClosedState :: [Party] -> NodeState SimpleTx

inClosedState' :: [Party] -> ConfirmedSnapshot SimpleTx -> NodeState SimpleTx

getConfirmedSnapshot :: NodeState tx -> Maybe (Snapshot tx)

data StepState tx

Constructors

StepState

Fields

nodeState :: NodeState tx

env :: Environment

ledger :: Ledger tx

runHeadLogic :: Monad m => Environment -> Ledger tx -> NodeState tx -> StateT (StepState tx) m a -> m a

getState :: MonadState (StepState tx) m => m (NodeState tx)

Retrieves the latest NodeState from within runHeadLogic.

step :: (MonadState (StepState tx) m, IsChainState tx, MonadTime m) => Input tx -> m (Outcome tx)

Calls update and aggregate to drive the runHeadLogic monad forward.

hasEffect :: (HasCallStack, IsChainState tx) => Outcome tx -> Effect tx -> IO ()

assertWait :: (HasCallStack, IsChainState tx) => Outcome tx -> WaitReason tx -> IO ()

hasEffectSatisfying :: (HasCallStack, IsChainState tx) => Outcome tx -> (Effect tx -> Bool) -> IO ()

hasNoEffectSatisfying :: (HasCallStack, IsChainState tx) => Outcome tx -> (Effect tx -> Bool) -> IO ()

hasStateChangedSatisfying :: (HasCallStack, IsChainState tx) => Outcome tx -> (StateChanged tx -> Bool) -> IO ()

hasNoStateChangedSatisfying :: (HasCallStack, IsChainState tx) => Outcome tx -> (StateChanged tx -> Bool) -> IO ()

testSnapshot :: Monoid (UTxOType tx) => SnapshotNumber -> SnapshotVersion -> [tx] -> UTxOType tx -> Snapshot tx

nowFromSlot :: (MonadFail m, MonadTime m) => ChainSlot -> m UTCTime

mkTimeHandleAt :: SlotNo -> UTCTime -> TimeHandle

Create a TimeHandle for a given slot and current wall-clock time. * Assumes 1-second slots (as in the devnet). * SystemStart is derived so that the given slot corresponds to the given time. * EraHistory is constructed to have a horizon sufficiently far in the future to accommodate time conversions.