Hydra.Model

Description

A Model of the Hydra head Protocol.

This model integrates in a single state-machine like abstraction the whole behaviour of a Hydra Head, taking into account both on-chain state and contracts, and off-chain interactions. It is written from the point of view of a pre-defined set of Hydra node operators that want to create a channel between them. It's a "happy path" model that does not implement any kind of adversarial behaviour and whose transactions are very simple: Each tx is a payment of one Ada-only UTxO transferred to another party in full, without any change.

More intricate and specialised models shall be developed once we get a firmer grasp of the whole framework, injecting faults, taking into account more parts of the stack, modelling more complex transactions schemes...

The Model

data WorldState

State maintained by the model.

Constructors

WorldState

Fields

hydraParties :: [(SigningKey HydraKey, CardanoSigningKey)]

List of parties identified by both signing keys required to run protocol. This list must not contain any duplicated key.

hydraState :: GlobalState

Expected consensus state All nodes should be in the same state.

availableToDeposit :: UTxOType Payment

UTxO available to be committed incrementally. NOTE: We must not add UTxO we decommitted to this as the Payment transaction model results in non-unique transaction ids when running the model.

data GlobalState

Global state of the Head protocol. While each participant in the Hydra Head protocol has its own private view of the state, we model the expected global state whose properties stem from the consensus built into the Head protocol. In other words, this state is what each node's local state should be eventually.

Constructors

Start
Start of the "world". This state is left implicit in the node's logic as it represents that state where the node does not even exist.

Idle

Fields

idleParties :: [Party]

cardanoKeys :: [VerificationKey PaymentKey]

contestationPeriod :: ContestationPeriod

toCommit :: Uncommitted

Initial

Fields

headIdVar :: Var HeadId

headParameters :: HeadParameters

commits :: Committed Payment

pendingCommits :: Uncommitted

Open

Fields

headIdVar :: Var HeadId

headParameters :: HeadParameters

offChainState :: OffChainState

committed :: Committed Payment

Closed

Fields

headParameters :: HeadParameters

closedUTxO :: UTxOType Payment

Final

Fields

finalUTxO :: UTxOType Payment

type Uncommitted = Map Party (UTxOType Payment)

newtype OffChainState

Constructors

OffChainState

Fields

confirmedUTxO :: UTxOType Payment

type ActualCommitted = UTxOType Payment

Generator Helper

genSeed :: Gen (Action WorldState ())

genToCommit :: (SigningKey HydraKey, CardanoSigningKey) -> Gen (Map Party [(CardanoSigningKey, Value)])

genContestationPeriod :: Gen ContestationPeriod

genInit :: [(SigningKey HydraKey, b)] -> Gen (Action WorldState HeadId)

genPayment :: WorldState -> Gen (Party, Payment)

unsafeConstructorName :: Show a => a -> String

partyKeys :: Gen [(SigningKey HydraKey, CardanoSigningKey)]

Generate a list of pairs of Hydra/Cardano signing keys. All the keys in this list are guaranteed to be unique.

Running the model

data Nodes m

Concrete state needed to run actions against the implementation. This state is used and might be updated when actually performing actions generated from the StateModel.

Constructors

Nodes

Fields

nodes :: Map Party (TestHydraClient Tx m)

Map from party identifiers to a handle for interacting with a node.

logger :: Tracer m (HydraLog Tx)

Logger used by each node. The reason we put this here is because the concrete value needs to be instantiated upon the test run initialisation, outiside of the model.

threads :: [Async m ()]

List of threads spawned when executing RunMonad

chain :: SimulatedChainNetwork Tx m

newtype RunState m

Constructors

RunState

Fields

nodesState :: TVar m (Nodes m)

newtype RunMonad m a

Our execution MonadTransformer.

This type is needed in order to keep the execution monad m abstract and thus simplify the definition of the RunModel instance which requires a proper definition of Realized type family. See this issue for a discussion on why this monad is needed.

We could perhaps getaway with it and just have a type based on IOSim monad but this is cumbersome to write.

Constructors

RunMonad

Fields

runMonad :: ReaderT (RunState m) m a

data RunException

Constructors

TransactionNotObserved Payment UTxO

UnexpectedParty Party

UnknownAddress AddressInEra [(AddressInEra, CardanoSigningKey)]

CannotFindSpendableUTxO Payment UTxO

sortTxOuts :: [TxOut ctx] -> [TxOut ctx]

Performing actions

testDepositPeriod :: DepositPeriod

Deposit period used by all nodes and the performDeposit.

seedWorld :: (MonadAsync m, MonadTimer m, MonadThrow (STM m), MonadLabelledSTM m, MonadFork m, MonadMask m, MonadDelay m, MonadTime m) => [(SigningKey HydraKey, CardanoSigningKey)] -> ContestationPeriod -> Uncommitted -> RunMonad m ()

performCommit :: (MonadThrow m, MonadTimer m, MonadAsync m, MonadLabelledSTM m) => HeadId -> Party -> [(CardanoSigningKey, Value)] -> RunMonad m ()

performDeposit :: (MonadThrow m, MonadTimer m, MonadAsync m, MonadTime m, MonadLabelledSTM m) => HeadId -> [(CardanoSigningKey, Value)] -> RunMonad m ()

performDecommit :: (MonadThrow m, MonadTimer m, MonadAsync m, MonadDelay m, MonadLabelledSTM m) => Party -> Payment -> RunMonad m ()

performNewTx :: (MonadThrow m, MonadAsync m, MonadTimer m, MonadDelay m, MonadLabelledSTM m) => Party -> Payment -> RunMonad m Payment

waitForOpen :: MonadDelay m => TestHydraClient tx m -> RunMonad m ()

Wait for the head to be open by searching from the beginning. Note that there rollbacks or multiple life-cycles of heads are not handled here.

waitForReadyToFanout :: MonadDelay m => TestHydraClient tx m -> RunMonad m ()

Wait for the head to be closed by searching from the beginning. Note that there rollbacks or multiple life-cycles of heads are not handled here.

sendsInput :: (MonadSTM m, MonadThrow m) => Party -> ClientInput Tx -> RunMonad m ()

getActorNode :: (MonadSTM m, MonadThrow m) => Party -> RunMonad m (TestHydraClient Tx m)

performInit :: (MonadThrow m, MonadAsync m, MonadTimer m, MonadLabelledSTM m) => Party -> RunMonad m HeadId

performAbort :: (MonadThrow m, MonadAsync m, MonadTimer m, MonadLabelledSTM m) => Party -> RunMonad m ()

performClose :: (MonadThrow m, MonadAsync m, MonadTimer m, MonadDelay m, MonadLabelledSTM m) => Party -> RunMonad m ()

performFanout :: (MonadThrow m, MonadAsync m, MonadDelay m) => Party -> RunMonad m UTxO

performCloseWithInitialSnapshot :: (MonadThrow m, MonadTimer m, MonadDelay m, MonadAsync m, MonadLabelledSTM m) => WorldState -> Party -> RunMonad m ()

performRollbackAndForward :: (MonadThrow m, MonadTimer m) => Natural -> RunMonad m ()

stopTheWorld :: MonadAsync m => RunMonad m ()

Utility functions

toTxOuts :: [(CardanoSigningKey, Value)] -> [TxOut CtxUTxO]

Convert payment-style utxos into transaction outputs.

toRealUTxO :: UTxOType Payment -> UTxOType Tx

Convert payment-style utxos into real utxos. The Payment tx domain is smaller than UTxO and we map every unique signer + value entry to a mocked TxIn on the real cardano domain.

mkTxOut :: CardanoSigningKey -> Value -> TxOut CtxUTxO

mkMockTxIn :: CardanoSigningKey -> Word -> TxIn

(===) :: (Eq a, Show a, Monad m) => a -> a -> PostconditionM m Bool

Like ===, but works in PostconditionM.

waitForUTxOToSpend :: forall m. MonadDelay m => UTxO -> CardanoSigningKey -> Value -> TestHydraClient Tx m -> m (Either UTxO (TxIn, TxOut CtxUTxO))

headUTxO :: (IsTx tx, MonadDelay m) => TestHydraClient tx m -> m (UTxOType tx)

isOwned :: CardanoSigningKey -> (TxIn, TxOut ctx) -> Bool

headIsOpen :: ServerOutput tx -> Bool

headIsReadyToFanout :: ServerOutput tx -> Bool