Hydra.Model.MockChain

Documentation

mockChainAndNetwork :: forall m. (MonadTimer m, MonadAsync m, MonadMask m, MonadThrow (STM m), MonadLabelledSTM m, MonadFork m, MonadDelay m) => Tracer m CardanoChainLog -> [(SigningKey HydraKey, CardanoSigningKey)] -> UTxO -> m (SimulatedChainNetwork Tx m)

Create a mocked chain which connects nodes through ChainSyncHandler and Chain interfaces. It calls connected chain sync handlers $sel:onRollForward:ChainSyncHandler on every blockTime and performs $sel:rollbackAndForward:SimulatedChainNetwork every couple blocks.

fixedTimeHandleIndefiniteHorizon :: Gen TimeHandle

Construct fixed TimeHandle that starts from 0 and has the era horizon far in the future. This is used in our Model tests and we want to make sure the tests finish before the horizon is reached to prevent the PastHorizon exceptions.

scriptLedger :: Ledger Tx

A trimmed down ledger whose only purpose is to validate on-chain scripts.

findOwnCardanoKey :: Party -> [(SigningKey HydraKey, CardanoSigningKey)] -> (VerificationKey PaymentKey, [VerificationKey PaymentKey])

Find Cardano vkey corresponding to our Hydra vkey using signing keys lookup. This is a bit cumbersome and a tribute to the fact the HydraNode itself has no direct knowledge of the cardano keys which are stored only at the ChainComponent level.

createMockNetwork :: MonadSTM m => DraftHydraNode Tx m -> TVar m [MockHydraNode m] -> Network m (Message Tx)

data MockHydraNode m

Constructors

MockHydraNode

Fields

node :: HydraNode Tx m

chainHandler :: ChainSyncHandler m

createMockChain :: (MonadTimer m, MonadThrow (STM m)) => Tracer m CardanoChainLog -> ChainContext -> SubmitTx m -> m TimeHandle -> TxIn -> LocalChainState m Tx -> Chain Tx m

flushQueue :: MonadSTM m => TQueue m a -> STM m [a]