Hydra.Network

Description

Asynchronous messaging interface to the Hydra Network, e.g to other Hydra nodes.

Concrete implementations are provided by submodules. Import those instead of this one if interested in actually configuring and running a real network layer.

Incoming and outgoing messages are modelled as Message data type.

Documentation

newtype Network m msg

Interface from the application to the network layer.

Constructors

Network

Fields

broadcast :: msg -> m ()

Send a msg to the whole configured hydra network including ourselves.

data Host

Constructors

Host

Fields

hostname :: Text

port :: PortNumber

newtype ProtocolVersion

Constructors

ProtocolVersion Natural

data NetworkCallback msg m

Interface from network layer to the application. XXX: Reliably delivering a message in the crash-recovery fault model is tricky. According to "Introduction to Reliable and Secure Distributed Programming" section "2.2.4 Crashes with recoveries" explains that storing to stable storage and just pointing to stored events is a better way.

Constructors

NetworkCallback

Fields

deliver :: msg -> m ()

The given msg was received from the network.

onConnectivity :: Connectivity -> m ()

The given Connectivity event was observed by network.

type NetworkComponent m inbound outbound a = NetworkCallback inbound m -> (Network m outbound -> m a) -> m a

A type tying both inbound and outbound messages sending in a single Component.

A NetworkComponent can have different inbound and outbound message types.

data Connectivity

Constructors

PeerConnected
Individual peer appeared alive on network.

Fields

peer :: Host

PeerDisconnected
Individual peer disappeared from network (has not been seen active in a while).

Fields

peer :: Host

NetworkConnected
Connected to Hydra network.

NetworkDisconnected
Disconnected from Hydra network.

VersionMismatch

Fields

ourVersion :: ProtocolVersion

theirVersion :: Maybe ProtocolVersion

ClusterIDMismatch

Fields

clusterPeers :: Text

data NetworkConfiguration

Configuration for a Node network layer.

Constructors

NetworkConfiguration

Fields

persistenceDir :: FilePath

Persistence directory

signingKey :: SigningKey HydraKey

This node's signing key. This is used to sign messages sent to peers.

otherParties :: [Party]

The list of peers Party known to this node.

listen :: Host

Address to listen on for incoming connections.

advertise :: Host

Address to advertise to peers.

peers :: [Host]

Addresses and ports of remote peers.

nodeId :: NodeId

This node's id.

whichEtcd :: WhichEtcd

Whether to use the system etcd (on the path) or the embedded one.

data WhichEtcd

Constructors

EmbeddedEtcd

SystemEtcd

newtype NodeId

Constructors

NodeId

Fields

nodeId :: Text

readHost :: MonadFail m => String -> m Host

readPort :: MonadFail m => String -> m PortNumber

showHost :: Host -> String

data IP

data PortNumber

Orphan instances

Arbitrary PortNumber

Instance details

Methods

arbitrary :: Gen PortNumber

shrink :: PortNumber -> [PortNumber]

FromJSON IP

Instance details

Methods

parseJSON :: Value -> Parser IP

parseJSONList :: Value -> Parser [IP]

omittedField :: Maybe IP

FromJSON PortNumber

Instance details

Methods

parseJSON :: Value -> Parser PortNumber

parseJSONList :: Value -> Parser [PortNumber]

omittedField :: Maybe PortNumber

ToJSON IP

Instance details

Methods

toJSON :: IP -> Value

toEncoding :: IP -> Encoding

toJSONList :: [IP] -> Value

toEncodingList :: [IP] -> Encoding

omitField :: IP -> Bool

ToJSON PortNumber

Instance details

Methods

toJSON :: PortNumber -> Value

toEncoding :: PortNumber -> Encoding

toJSONList :: [PortNumber] -> Value

toEncodingList :: [PortNumber] -> Encoding

omitField :: PortNumber -> Bool