Hydra.Tx.Contract.Increment

Documentation

healthyIncrementTx :: (Tx, UTxO)

healthyDepositInput :: TxIn

healthyDeposited :: UTxO

somePartyCardanoVerificationKey :: VerificationKey PaymentKey

healthySigningKeys :: [SigningKey HydraKey]

healthyParticipants :: [VerificationKey PaymentKey]

healthyParties :: [Party]

healthyOnChainParties :: [Party]

healthySignature :: MultiSignature (Snapshot Tx)

healthySnapshotNumber :: SnapshotNumber

healthySnapshotVersion :: SnapshotVersion

healthySnapshot :: Snapshot Tx

healthyContestationPeriod :: ContestationPeriod

healthyUTxO :: UTxO

healthyDatum :: State

data IncrementMutation

Constructors

DepositMutateDepositPeriod
Move the deadline from the deposit datum back in time so that the increment upper bound is after the deadline

DepositMutateHeadId
Alter the head id

IncrementMutateParties
Change parties in increment output datum

IncrementUseDifferentSnapshotVersion
New version is incremented correctly

ProduceInvalidSignatures
Produce invalid signatures

ChangeHeadValue
Change the head value

AlterRequiredSigner
Change the required signers

IncrementDifferentClaimRedeemer
Alter the Claim redeemer TxOutRef

genIncrementMutation :: (Tx, UTxO) -> Gen SomeMutation