Hydra.Tx.Fanout

Creation

fanoutTx

Arguments

:: ScriptRegistry
Published Hydra scripts to reference.

-> UTxO
Snapshotted UTxO to fanout on layer 1

-> Maybe UTxO
Snapshotted commit UTxO to fanout on layer 1

-> Maybe UTxO
Snapshotted decommit UTxO to fanout on layer 1

-> (TxIn, TxOut CtxUTxO)
Everything needed to spend the Head state-machine output.

-> SlotNo
Contestation deadline as SlotNo, used to set lower tx validity bound.

-> PlutusScript
Minting Policy script, made from initial seed

-> Tx

Create the fanout transaction, which distributes the closed state accordingly. The head validator allows fanout only > deadline, so we need to set the lower bound to be deadline + 1 slot.

Observation

data FanoutObservation

Constructors

FanoutObservation

Fields

headId :: HeadId

fanoutUTxO :: UTxO

observeFanoutTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe FanoutObservation

Identify a fanout tx by lookup up the input spending the Head output and decoding its redeemer.