HydraNode

Client to interact with a hydra-node

data HydraClient

Constructors

HydraClient

Fields

hydraNodeId :: Int

apiHost :: Host

connection :: Connection

tracer :: Tracer IO HydraNodeLog

input :: Text -> [Pair] -> Value

Create an input as expected by send.

send :: HydraClient -> Value -> IO ()

waitNext :: HasCallStack => HydraClient -> IO Value

output :: Text -> [Pair] -> Value

Create an output as expected by waitFor and waitForAll.

setupBFDelay :: NominalDiffTime -> IO NominalDiffTime

waitFor :: HasCallStack => Tracer IO HydraNodeLog -> NominalDiffTime -> [HydraClient] -> Value -> IO ()

Wait some time for a single API server output from each of given nodes. This function waits for delay seconds for message expected to be seen by all given nodes.

waitNoMatch :: HasCallStack => NominalDiffTime -> HydraClient -> (Value -> Maybe a) -> IO ()

Wait up to some time and succeed if no API server output matches the given predicate.

waitMatch :: HasCallStack => NominalDiffTime -> HydraClient -> (Value -> Maybe a) -> IO a

Wait up to some time for an API server output to match the given predicate.

waitForAllMatch :: (Eq a, Show a, HasCallStack) => NominalDiffTime -> [HydraClient] -> (Value -> Maybe a) -> IO a

Wait up to some delay for some JSON Value to match given function.

This is a generalisation of waitMatch to multiple nodes.

waitForAll :: HasCallStack => Tracer IO HydraNodeLog -> NominalDiffTime -> [HydraClient] -> [Value] -> IO ()

Wait some time for a list of outputs from each of given nodes. This function is the generalised version of waitFor, allowing several messages to be waited for and received in any order.

requestCommitTx :: HydraClient -> UTxO -> IO Tx

Helper to make it easy to obtain a commit tx using some wallet utxo

postDecommit :: HydraClient -> Tx -> IO ()

Submit a decommit transaction to the hydra-node.

getProtocolParameters :: HydraClient -> IO (PParams LedgerEra)

Get the protocol-parameters from the hydra-node.

getSnapshotUTxO :: HydraClient -> IO UTxO

Get the latest snapshot UTxO from the hydra-node. NOTE: While we usually avoid parsing responses using the same data types as the system under test, this parses the response as a UTxO type as we often need to pick it apart.

getSnapshotConfirmed :: HydraClient -> IO (ConfirmedSnapshot Tx)

Get the latest snapshot from the hydra-node. NOTE: While we usually avoid parsing responses using the same data types as the system under test, this parses the response as a ConfirmedSnapshot type as we often need to pick it apart.

getSnapshotLastSeen :: HydraClient -> IO (SeenSnapshot Tx)

Get the latest seen snapshot from the hydra-node.

getMetrics :: HasCallStack => HydraClient -> IO ByteString

Start / connect to a cluster of nodes

withHydraCluster

Arguments

:: HasCallStack

=> Tracer IO HydraNodeLog

-> FilePath

-> SocketPath

-> Int
First node id This sets the starting point for assigning ports

-> [(VerificationKey PaymentKey, SigningKey PaymentKey)]
NOTE: This decides on the size of the cluster!

-> [SigningKey HydraKey]

-> [TxId]
Transaction ids at which Hydra scripts should have been published.

-> ContestationPeriod

-> (NonEmpty HydraClient -> IO a)

-> IO a

Start / connect to a hydra-node

preparePParams :: ChainConfig -> FilePath -> (Value -> Value) -> IO FilePath

Prepare protocol-parameters to run a hydra-node with given ChainConfig and using the config from config/.

prepareHydraNode :: HasCallStack => ChainConfig -> FilePath -> Int -> SigningKey HydraKey -> [VerificationKey HydraKey] -> [Int] -> (Value -> Value) -> IO RunOptions

Prepare RunOptions to run a hydra-node with given ChainConfig and using the config from config/.

withPreparedHydraNodeInSync :: HasCallStack => Tracer IO HydraNodeLog -> FilePath -> Int -> RunOptions -> (HydraClient -> IO a) -> IO a

Run a hydra-node with given RunOptions and in sync with chain backend.

withPreparedHydraNode :: HasCallStack => Tracer IO HydraNodeLog -> FilePath -> Int -> RunOptions -> (HydraClient -> IO a) -> IO a

Run a hydra-node with given RunOptions.

withHydraNode :: HasCallStack => Tracer IO HydraNodeLog -> ChainConfig -> FilePath -> Int -> SigningKey HydraKey -> [VerificationKey HydraKey] -> [Int] -> (HydraClient -> IO a) -> IO a

Run a hydra-node with given ChainConfig and using the config from config/.

withHydraNodeCatchingUp :: HasCallStack => Tracer IO HydraNodeLog -> ChainConfig -> FilePath -> Int -> SigningKey HydraKey -> [VerificationKey HydraKey] -> [Int] -> (HydraClient -> IO a) -> IO a

Run a hydra-node with given ChainConfig and using the config from config and catching up with chain backend/.

withConnectionToNode :: forall a. Tracer IO HydraNodeLog -> Int -> (HydraClient -> IO a) -> IO a

withConnectionToNodeHost :: forall a. Tracer IO HydraNodeLog -> Int -> Host -> Maybe String -> (HydraClient -> IO a) -> IO a

waitForNodesConnected :: Tracer IO HydraNodeLog -> NominalDiffTime -> NonEmpty HydraClient -> IO ()

waitForNodesDisconnected :: Tracer IO HydraNodeLog -> NominalDiffTime -> NonEmpty HydraClient -> IO ()

waitForNodesSynced :: Tracer IO HydraNodeLog -> NominalDiffTime -> NonEmpty HydraClient -> IO ()

data HydraNodeLog

Constructors

HydraNodeCommandSpec

Fields

cmd :: Text

NodeStarted

Fields

nodeId :: Int

SentMessage

Fields

nodeId :: Int

message :: Value

StartWaiting

Fields

nodeIds :: [Int]

messages :: [Value]

ReceivedMessage

Fields

nodeId :: Int

message :: Value

EndWaiting

Fields

nodeId :: Int